refactor(GlpiRestClient): update sample codes and unit tests
