+++ lib/GlpiRestClient.js
@@ -1,1 +1,1 @@
-!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t(require("https"),require("url")):"function"==typeof define&&define.amd?define(["https","url"],t):"object"==typeof exports?exports.GlpiRestClient=t(require("https"),require("url")):e.GlpiRestClient=t(e.https,e.url)}("undefined"!=typeof self?self:this,function(e,t){return function(e){var t={};function r(n){if(t[n])return t[n].exports;var o=t[n]={i:n,l:!1,exports:{}};return e[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}return r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{configurable:!1,enumerable:!0,get:n})},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=0)}([function(e,t,r){"use strict";(function(t){var n=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e},o=function(){function e(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}return function(t,r,n){return r&&e(t.prototype,r),n&&e(t,n),t}}();var i=r(6),a=r(7).URL,s=r(8),u=function(){function e(t){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),this._url=t,this._sessionToken="",this._appToken=""}return o(e,[{key:"_prepareRequest",value:function(e,t,r,o){var s=new a(this._url);s.pathname+="/"+t;var u={"Content-Type":"application/json","Session-Token":this.sessionToken,"App-Token":this.appToken},c="";if(r&&(r.headers&&(u=n({},u,r.headers)),r["App-Token"]&&(this.appToken=r["App-Token"]),r.input&&(c=JSON.stringify(r),u=n({},u,{"Content-Length":c.length})),r.profiles_id&&(c=JSON.stringify(r),u=n({},u,{"Content-Length":c.length})),r.activeEntities&&(c=JSON.stringify(r.activeEntities),u=n({},u,{"Content-Length":c.length})),r.lostPassword&&(c=JSON.stringify(r.lostPassword),u=n({},u,{"Content-Length":c.length})),r.queryString))for(var l in r.queryString)r.queryString[l]&&s.searchParams.append(l,r.queryString[l]);var m={hostname:s.hostname,port:443,path:s.pathname+s.search,method:e,headers:u},f=i.request(m,function(e){var t="";e.on("data",function(e){t+=e}),e.on("end",function(){o(e,t)})});return f.on("error",function(){}),f.write(c),f}},{key:"_requestException",value:function(e,t){return{status:e,data:n({},t)}}},{key:"_assessStatus",value:function(e,t){if(e>=400&&e<600)return e>=400&&e<500?this._requestException(e,{codeError:JSON.parse(t)[0],messageError:JSON.parse(t)[1]}):this._requestException(e,JSON.parse(t))}},{key:"_isNumber",value:function(e){return!isNaN(e-0)&&null!==e&&""!==e&&!1!==e}},{key:"initSessionByCredentials",value:function(e,r,o){var i=this;return new Promise(function(a,s){try{var u={headers:{Authorization:"Basic "+t.from(e+":"+r).toString("base64")}};o&&(u.headers=n({},u.headers,{"App-Token":o})),i._prepareRequest("GET","initSession",u,function(e,t){var r=i._assessStatus(e.statusCode,t);r?s(r):(i.sessionToken=JSON.parse(t).session_token,a({status:e.statusCode,data:JSON.parse(t)}))}).end()}catch(e){s(e)}})}},{key:"initSessionByUserToken",value:function(e,t){var r=this;return new Promise(function(o,i){try{var a={headers:{Authorization:"user_token "+e}};t&&(a.headers=n({},a.headers,{"App-Token":t})),r._prepareRequest("GET","initSession",a,function(e,t){var n=r._assessStatus(e.statusCode,t);n?i(n):(r.sessionToken=JSON.parse(t).session_token,o({status:e.statusCode,data:JSON.parse(t)}))}).end()}catch(e){i(e)}})}},{key:"killSession",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","killSession",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):(e.sessionToken="",t({status:n.statusCode,data:{message:"User logout successfully"}}))}).end()}catch(e){r(e)}})}},{key:"getMyProfiles",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getMyProfiles",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o).myprofiles})}).end()}catch(e){r(e)}})}},{key:"getActiveProfile",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getActiveProfile",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o).active_profile})}).end()}catch(e){r(e)}})}},{key:"getMyEntities",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getMyEntities",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o).myentities})}).end()}catch(e){r(e)}})}},{key:"getActiveEntities",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getActiveEntities",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o)})}).end()}catch(e){r(e)}})}},{key:"getFullSession",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getFullSession",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o)})}).end()}catch(e){r(e)}})}},{key:"getGlpiConfig",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getGlpiConfig",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o)})}).end()}catch(e){r(e)}})}},{key:"getAllItems",value:function(e,t){var r=this;return new Promise(function(n,o){try{if(!e)throw r._requestException(400,"Invalid itemtype");if(e!==s[e.name])throw r._requestException(400,"Invalid itemtype");var i=null;t&&(i={queryString:t}),r._prepareRequest("GET",e.name,i,function(e,t){var i=r._assessStatus(e.statusCode,t);i?o(i):n({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){o(e)}})}},{key:"getAnItem",value:function(e,t,r){var n=this;return new Promise(function(o,i){try{if(!e)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(!n._isNumber(t))throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});var a=null;r&&(a={queryString:r});var u=e.name+"/"+t;n._prepareRequest("GET",u,a,function(e,t){var r=n._assessStatus(e.statusCode,t);r?i(r):o({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){i(e)}})}},{key:"getSubItems",value:function(e,t,r,n){var o=this;return new Promise(function(i,a){try{if(!e)throw o._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw o._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(!r)throw o._requestException(400,{errorCode:"",errorMessage:"Invalid sub itemtype"});if(r!==s[r.name])throw o._requestException(400,{errorCode:"",errorMessage:"Invalid sub itemtype"});if(!o._isNumber(t))throw o._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw o._requestException(400,{errorCode:"",errorMessage:"Invalid id"});var u=null;n&&(u={queryString:n});var c=e.name+"/"+t+"/"+r.name;o._prepareRequest("GET",c,u,function(e,t){var r=o._assessStatus(e.statusCode,t);r?a(r):i({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){a(e)}})}},{key:"addItem",value:function(e,t){var r=this;return new Promise(function(n,o){try{if(!e)throw r._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw r._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});var i={input:t},a=e.name;r._prepareRequest("POST",a,i,function(e,t){var i=r._assessStatus(e.statusCode,t);i?o(i):n({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){o(e)}})}},{key:"deleteItem",value:function(e,t,r,o){var i=this;return new Promise(function(a,u){try{if(!e)throw i._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw i._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});var c={};o&&(c={queryString:o});var l=e.name;if(t){if(!i._isNumber(t))throw i._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw i._requestException(400,{errorCode:"",errorMessage:"Invalid id"});l+="/"+t}else{if(!r)throw i._requestException(400,{errorCode:"",errorMessage:"Invalid input"});c=n({},c,{input:r})}i._prepareRequest("DELETE",l,c,function(e,t){var r=i._assessStatus(e.statusCode,t);r?u(r):a({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){u(e)}})}},{key:"updateItem",value:function(e,t,r){var n=this;return new Promise(function(o,i){try{if(!e)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});var a="",u=e.name;if(t){if(!n._isNumber(t))throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});u+="/"+t}else{if(!r)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid input"});a={input:r}}n._prepareRequest("PUT",u,a,function(e,t){var r=n._assessStatus(e.statusCode,t);r?i(r):o({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){i(e)}})}},{key:"changeActiveProfile",value:function(e){var t=this;return new Promise(function(r,n){try{if(!t._isNumber(e))throw t._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(e<0)throw t._requestException(400,{errorCode:"",errorMessage:"Invalid id"});var o={profiles_id:e};t._prepareRequest("POTS","changeActiveProfile",o,function(e,o){var i=t._assessStatus(e.statusCode,o);i?n(i):r({status:e.statusCode,data:{message:"Active profile successfully changed"}})}).end()}catch(e){n(e)}})}},{key:"changeActiveEntities",value:function(e,t){var r=this;return new Promise(function(o,i){try{var a={activeEntities:{}};if(e){if(!r._isNumber(e))throw r._requestException(400,{errorCode:"",errorMessage:"Invalid entities_id"});if(e<0)throw r._requestException(400,{errorCode:"",errorMessage:"Invalid entities_id"});a={activeEntities:e}}if(t){if("boolean"!=typeof t)throw r._requestException(400,{errorCode:"",errorMessage:"Invalid is_recursive"});a={activeEntities:n({},a.activeEntities,{is_recursive:t})}}r._prepareRequest("POTS","changeActiveEntities",a,function(e,t){var n=r._assessStatus(e.statusCode,t);n?i(n):o(t?{status:e.statusCode,data:{message:"Active entities successfully changed"}}:{status:e.statusCode,data:{message:"Failed to change active entities"}})}).end()}catch(e){i(e)}})}},{key:"resetPasswordRequest",value:function(e){var t=this;return new Promise(function(r,n){try{var o={lostPassword:{email:e}};t._prepareRequest("PUT","lostPassword",o,function(e,o){var i=t._assessStatus(e.statusCode,o);i?n(i):r({status:e.statusCode,data:JSON.parse(o)[0]})}).end()}catch(e){n(e)}})}},{key:"passwordReset",value:function(e,t,r){var n=this;return new Promise(function(o,i){try{var a={lostPassword:{email:e,password_forget_token:t,password:r}};n._prepareRequest("PUT","lostPassword",a,function(e,t){var r=n._assessStatus(e.statusCode,t);r?i(r):o({status:e.statusCode,data:JSON.parse(t)[0]})}).end()}catch(e){i(e)}})}},{key:"url",get:function(){return this._url},set:function(e){this._url=e}},{key:"sessionToken",get:function(){return this._sessionToken},set:function(e){e&&(this._sessionToken=e)}},{key:"appToken",get:function(){return this._appToken},set:function(e){e&&(this._appToken=e)}}]),e}();e.exports=u}).call(t,r(1).Buffer)},function(e,t,r){"use strict";(function(e){var n=r(3),o=r(4),i=r(5);function a(){return u.TYPED_ARRAY_SUPPORT?2147483647:1073741823}function s(e,t){if(a()<t)throw new RangeError("Invalid typed array length");return u.TYPED_ARRAY_SUPPORT?(e=new Uint8Array(t)).__proto__=u.prototype:(null===e&&(e=new u(t)),e.length=t),e}function u(e,t,r){if(!(u.TYPED_ARRAY_SUPPORT||this instanceof u))return new u(e,t,r);if("number"==typeof e){if("string"==typeof t)throw new Error("If encoding is specified then the first argument must be a string");return m(this,e)}return c(this,e,t,r)}function c(e,t,r,n){if("number"==typeof t)throw new TypeError('"value" argument must not be a number');return"undefined"!=typeof ArrayBuffer&&t instanceof ArrayBuffer?function(e,t,r,n){if(t.byteLength,r<0||t.byteLength<r)throw new RangeError("'offset' is out of bounds");if(t.byteLength<r+(n||0))throw new RangeError("'length' is out of bounds");t=void 0===r&&void 0===n?new Uint8Array(t):void 0===n?new Uint8Array(t,r):new Uint8Array(t,r,n);u.TYPED_ARRAY_SUPPORT?(e=t).__proto__=u.prototype:e=f(e,t);return e}(e,t,r,n):"string"==typeof t?function(e,t,r){"string"==typeof r&&""!==r||(r="utf8");if(!u.isEncoding(r))throw new TypeError('"encoding" must be a valid string encoding');var n=0|h(t,r),o=(e=s(e,n)).write(t,r);o!==n&&(e=e.slice(0,o));return e}(e,t,r):function(e,t){if(u.isBuffer(t)){var r=0|p(t.length);return 0===(e=s(e,r)).length?e:(t.copy(e,0,0,r),e)}if(t){if("undefined"!=typeof ArrayBuffer&&t.buffer instanceof ArrayBuffer||"length"in t)return"number"!=typeof t.length||(n=t.length)!=n?s(e,0):f(e,t);if("Buffer"===t.type&&i(t.data))return f(e,t.data)}var n;throw new TypeError("First argument must be a string, Buffer, ArrayBuffer, Array, or array-like object.")}(e,t)}function l(e){if("number"!=typeof e)throw new TypeError('"size" argument must be a number');if(e<0)throw new RangeError('"size" argument must not be negative')}function m(e,t){if(l(t),e=s(e,t<0?0:0|p(t)),!u.TYPED_ARRAY_SUPPORT)for(var r=0;r<t;++r)e[r]=0;return e}function f(e,t){var r=t.length<0?0:0|p(t.length);e=s(e,r);for(var n=0;n<r;n+=1)e[n]=255&t[n];return e}function p(e){if(e>=a())throw new RangeError("Attempt to allocate Buffer larger than maximum size: 0x"+a().toString(16)+" bytes");return 0|e}function h(e,t){if(u.isBuffer(e))return e.length;if("undefined"!=typeof ArrayBuffer&&"function"==typeof ArrayBuffer.isView&&(ArrayBuffer.isView(e)||e instanceof ArrayBuffer))return e.byteLength;"string"!=typeof e&&(e=""+e);var r=e.length;if(0===r)return 0;for(var n=!1;;)switch(t){case"ascii":case"latin1":case"binary":return r;case"utf8":case"utf-8":case void 0:return O(e).length;case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return 2*r;case"hex":return r>>>1;case"base64":return L(e).length;default:if(n)return O(e).length;t=(""+t).toLowerCase(),n=!0}}function y(e,t,r){var n=e[t];e[t]=e[r],e[r]=n}function d(e,t,r,n,o){if(0===e.length)return-1;if("string"==typeof r?(n=r,r=0):r>2147483647?r=2147483647:r<-2147483648&&(r=-2147483648),r=+r,isNaN(r)&&(r=o?0:e.length-1),r<0&&(r=e.length+r),r>=e.length){if(o)return-1;r=e.length-1}else if(r<0){if(!o)return-1;r=0}if("string"==typeof t&&(t=u.from(t,n)),u.isBuffer(t))return 0===t.length?-1:g(e,t,r,n,o);if("number"==typeof t)return t&=255,u.TYPED_ARRAY_SUPPORT&&"function"==typeof Uint8Array.prototype.indexOf?o?Uint8Array.prototype.indexOf.call(e,t,r):Uint8Array.prototype.lastIndexOf.call(e,t,r):g(e,[t],r,n,o);throw new TypeError("val must be string, number or Buffer")}function g(e,t,r,n,o){var i,a=1,s=e.length,u=t.length;if(void 0!==n&&("ucs2"===(n=String(n).toLowerCase())||"ucs-2"===n||"utf16le"===n||"utf-16le"===n)){if(e.length<2||t.length<2)return-1;a=2,s/=2,u/=2,r/=2}function c(e,t){return 1===a?e[t]:e.readUInt16BE(t*a)}if(o){var l=-1;for(i=r;i<s;i++)if(c(e,i)===c(t,-1===l?0:i-l)){if(-1===l&&(l=i),i-l+1===u)return l*a}else-1!==l&&(i-=i-l),l=-1}else for(r+u>s&&(r=s-u),i=r;i>=0;i--){for(var m=!0,f=0;f<u;f++)if(c(e,i+f)!==c(t,f)){m=!1;break}if(m)return i}return-1}function C(e,t,r,n){r=Number(r)||0;var o=e.length-r;n?(n=Number(n))>o&&(n=o):n=o;var i=t.length;if(i%2!=0)throw new TypeError("Invalid hex string");n>i/2&&(n=i/2);for(var a=0;a<n;++a){var s=parseInt(t.substr(2*a,2),16);if(isNaN(s))return a;e[r+a]=s}return a}function v(e,t,r,n){return q(function(e){for(var t=[],r=0;r<e.length;++r)t.push(255&e.charCodeAt(r));return t}(t),e,r,n)}function w(e,t,r){return 0===t&&r===e.length?n.fromByteArray(e):n.fromByteArray(e.slice(t,r))}function T(e,t,r){r=Math.min(e.length,r);for(var n=[],o=t;o<r;){var i,a,s,u,c=e[o],l=null,m=c>239?4:c>223?3:c>191?2:1;if(o+m<=r)switch(m){case 1:c<128&&(l=c);break;case 2:128==(192&(i=e[o+1]))&&(u=(31&c)<<6|63&i)>127&&(l=u);break;case 3:i=e[o+1],a=e[o+2],128==(192&i)&&128==(192&a)&&(u=(15&c)<<12|(63&i)<<6|63&a)>2047&&(u<55296||u>57343)&&(l=u);break;case 4:i=e[o+1],a=e[o+2],s=e[o+3],128==(192&i)&&128==(192&a)&&128==(192&s)&&(u=(15&c)<<18|(63&i)<<12|(63&a)<<6|63&s)>65535&&u<1114112&&(l=u)}null===l?(l=65533,m=1):l>65535&&(l-=65536,n.push(l>>>10&1023|55296),l=56320|1023&l),n.push(l),o+=m}return function(e){var t=e.length;if(t<=_)return String.fromCharCode.apply(String,e);var r="",n=0;for(;n<t;)r+=String.fromCharCode.apply(String,e.slice(n,n+=_));return r}(n)}t.Buffer=u,t.SlowBuffer=function(e){+e!=e&&(e=0);return u.alloc(+e)},t.INSPECT_MAX_BYTES=50,u.TYPED_ARRAY_SUPPORT=void 0!==e.TYPED_ARRAY_SUPPORT?e.TYPED_ARRAY_SUPPORT:function(){try{var e=new Uint8Array(1);return e.__proto__={__proto__:Uint8Array.prototype,foo:function(){return 42}},42===e.foo()&&"function"==typeof e.subarray&&0===e.subarray(1,1).byteLength}catch(e){return!1}}(),t.kMaxLength=a(),u.poolSize=8192,u._augment=function(e){return e.__proto__=u.prototype,e},u.from=function(e,t,r){return c(null,e,t,r)},u.TYPED_ARRAY_SUPPORT&&(u.prototype.__proto__=Uint8Array.prototype,u.__proto__=Uint8Array,"undefined"!=typeof Symbol&&Symbol.species&&u[Symbol.species]===u&&Object.defineProperty(u,Symbol.species,{value:null,configurable:!0})),u.alloc=function(e,t,r){return n=null,i=t,a=r,l(o=e),o<=0?s(n,o):void 0!==i?"string"==typeof a?s(n,o).fill(i,a):s(n,o).fill(i):s(n,o);var n,o,i,a},u.allocUnsafe=function(e){return m(null,e)},u.allocUnsafeSlow=function(e){return m(null,e)},u.isBuffer=function(e){return!(null==e||!e._isBuffer)},u.compare=function(e,t){if(!u.isBuffer(e)||!u.isBuffer(t))throw new TypeError("Arguments must be Buffers");if(e===t)return 0;for(var r=e.length,n=t.length,o=0,i=Math.min(r,n);o<i;++o)if(e[o]!==t[o]){r=e[o],n=t[o];break}return r<n?-1:n<r?1:0},u.isEncoding=function(e){switch(String(e).toLowerCase()){case"hex":case"utf8":case"utf-8":case"ascii":case"latin1":case"binary":case"base64":case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return!0;default:return!1}},u.concat=function(e,t){if(!i(e))throw new TypeError('"list" argument must be an Array of Buffers');if(0===e.length)return u.alloc(0);var r;if(void 0===t)for(t=0,r=0;r<e.length;++r)t+=e[r].length;var n=u.allocUnsafe(t),o=0;for(r=0;r<e.length;++r){var a=e[r];if(!u.isBuffer(a))throw new TypeError('"list" argument must be an Array of Buffers');a.copy(n,o),o+=a.length}return n},u.byteLength=h,u.prototype._isBuffer=!0,u.prototype.swap16=function(){var e=this.length;if(e%2!=0)throw new RangeError("Buffer size must be a multiple of 16-bits");for(var t=0;t<e;t+=2)y(this,t,t+1);return this},u.prototype.swap32=function(){var e=this.length;if(e%4!=0)throw new RangeError("Buffer size must be a multiple of 32-bits");for(var t=0;t<e;t+=4)y(this,t,t+3),y(this,t+1,t+2);return this},u.prototype.swap64=function(){var e=this.length;if(e%8!=0)throw new RangeError("Buffer size must be a multiple of 64-bits");for(var t=0;t<e;t+=8)y(this,t,t+7),y(this,t+1,t+6),y(this,t+2,t+5),y(this,t+3,t+4);return this},u.prototype.toString=function(){var e=0|this.length;return 0===e?"":0===arguments.length?T(this,0,e):function(e,t,r){var n=!1;if((void 0===t||t<0)&&(t=0),t>this.length)return"";if((void 0===r||r>this.length)&&(r=this.length),r<=0)return"";if((r>>>=0)<=(t>>>=0))return"";for(e||(e="utf8");;)switch(e){case"hex":return S(this,t,r);case"utf8":case"utf-8":return T(this,t,r);case"ascii":return P(this,t,r);case"latin1":case"binary":return R(this,t,r);case"base64":return w(this,t,r);case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return k(this,t,r);default:if(n)throw new TypeError("Unknown encoding: "+e);e=(e+"").toLowerCase(),n=!0}}.apply(this,arguments)},u.prototype.equals=function(e){if(!u.isBuffer(e))throw new TypeError("Argument must be a Buffer");return this===e||0===u.compare(this,e)},u.prototype.inspect=function(){var e="",r=t.INSPECT_MAX_BYTES;return this.length>0&&(e=this.toString("hex",0,r).match(/.{2}/g).join(" "),this.length>r&&(e+=" ... ")),"<Buffer "+e+">"},u.prototype.compare=function(e,t,r,n,o){if(!u.isBuffer(e))throw new TypeError("Argument must be a Buffer");if(void 0===t&&(t=0),void 0===r&&(r=e?e.length:0),void 0===n&&(n=0),void 0===o&&(o=this.length),t<0||r>e.length||n<0||o>this.length)throw new RangeError("out of range index");if(n>=o&&t>=r)return 0;if(n>=o)return-1;if(t>=r)return 1;if(t>>>=0,r>>>=0,n>>>=0,o>>>=0,this===e)return 0;for(var i=o-n,a=r-t,s=Math.min(i,a),c=this.slice(n,o),l=e.slice(t,r),m=0;m<s;++m)if(c[m]!==l[m]){i=c[m],a=l[m];break}return i<a?-1:a<i?1:0},u.prototype.includes=function(e,t,r){return-1!==this.indexOf(e,t,r)},u.prototype.indexOf=function(e,t,r){return d(this,e,t,r,!0)},u.prototype.lastIndexOf=function(e,t,r){return d(this,e,t,r,!1)},u.prototype.write=function(e,t,r,n){if(void 0===t)n="utf8",r=this.length,t=0;else if(void 0===r&&"string"==typeof t)n=t,r=this.length,t=0;else{if(!isFinite(t))throw new Error("Buffer.write(string, encoding, offset[, length]) is no longer supported");t|=0,isFinite(r)?(r|=0,void 0===n&&(n="utf8")):(n=r,r=void 0)}var o=this.length-t;if((void 0===r||r>o)&&(r=o),e.length>0&&(r<0||t<0)||t>this.length)throw new RangeError("Attempt to write outside buffer bounds");n||(n="utf8");for(var i,a,s,u,c,l,m,f,p,h=!1;;)switch(n){case"hex":return C(this,e,t,r);case"utf8":case"utf-8":return f=t,p=r,q(O(e,(m=this).length-f),m,f,p);case"ascii":return v(this,e,t,r);case"latin1":case"binary":return v(this,e,t,r);case"base64":return u=this,c=t,l=r,q(L(e),u,c,l);case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return a=t,s=r,q(function(e,t){for(var r,n,o,i=[],a=0;a<e.length&&!((t-=2)<0);++a)r=e.charCodeAt(a),n=r>>8,o=r%256,i.push(o),i.push(n);return i}(e,(i=this).length-a),i,a,s);default:if(h)throw new TypeError("Unknown encoding: "+n);n=(""+n).toLowerCase(),h=!0}},u.prototype.toJSON=function(){return{type:"Buffer",data:Array.prototype.slice.call(this._arr||this,0)}};var _=4096;function P(e,t,r){var n="";r=Math.min(e.length,r);for(var o=t;o<r;++o)n+=String.fromCharCode(127&e[o]);return n}function R(e,t,r){var n="";r=Math.min(e.length,r);for(var o=t;o<r;++o)n+=String.fromCharCode(e[o]);return n}function S(e,t,r){var n=e.length;(!t||t<0)&&(t=0),(!r||r<0||r>n)&&(r=n);for(var o="",i=t;i<r;++i)o+=B(e[i]);return o}function k(e,t,r){for(var n=e.slice(t,r),o="",i=0;i<n.length;i+=2)o+=String.fromCharCode(n[i]+256*n[i+1]);return o}function D(e,t,r){if(e%1!=0||e<0)throw new RangeError("offset is not uint");if(e+t>r)throw new RangeError("Trying to access beyond buffer length")}function I(e,t,r,n,o,i){if(!u.isBuffer(e))throw new TypeError('"buffer" argument must be a Buffer instance');if(t>o||t<i)throw new RangeError('"value" argument is out of bounds');if(r+n>e.length)throw new RangeError("Index out of range")}function E(e,t,r,n){t<0&&(t=65535+t+1);for(var o=0,i=Math.min(e.length-r,2);o<i;++o)e[r+o]=(t&255<<8*(n?o:1-o))>>>8*(n?o:1-o)}function b(e,t,r,n){t<0&&(t=4294967295+t+1);for(var o=0,i=Math.min(e.length-r,4);o<i;++o)e[r+o]=t>>>8*(n?o:3-o)&255}function N(e,t,r,n,o,i){if(r+n>e.length)throw new RangeError("Index out of range");if(r<0)throw new RangeError("Index out of range")}function M(e,t,r,n,i){return i||N(e,0,r,4),o.write(e,t,r,n,23,4),r+4}function A(e,t,r,n,i){return i||N(e,0,r,8),o.write(e,t,r,n,52,8),r+8}u.prototype.slice=function(e,t){var r,n=this.length;if(e=~~e,t=void 0===t?n:~~t,e<0?(e+=n)<0&&(e=0):e>n&&(e=n),t<0?(t+=n)<0&&(t=0):t>n&&(t=n),t<e&&(t=e),u.TYPED_ARRAY_SUPPORT)(r=this.subarray(e,t)).__proto__=u.prototype;else{var o=t-e;r=new u(o,void 0);for(var i=0;i<o;++i)r[i]=this[i+e]}return r},u.prototype.readUIntLE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=this[e],o=1,i=0;++i<t&&(o*=256);)n+=this[e+i]*o;return n},u.prototype.readUIntBE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=this[e+--t],o=1;t>0&&(o*=256);)n+=this[e+--t]*o;return n},u.prototype.readUInt8=function(e,t){return t||D(e,1,this.length),this[e]},u.prototype.readUInt16LE=function(e,t){return t||D(e,2,this.length),this[e]|this[e+1]<<8},u.prototype.readUInt16BE=function(e,t){return t||D(e,2,this.length),this[e]<<8|this[e+1]},u.prototype.readUInt32LE=function(e,t){return t||D(e,4,this.length),(this[e]|this[e+1]<<8|this[e+2]<<16)+16777216*this[e+3]},u.prototype.readUInt32BE=function(e,t){return t||D(e,4,this.length),16777216*this[e]+(this[e+1]<<16|this[e+2]<<8|this[e+3])},u.prototype.readIntLE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=this[e],o=1,i=0;++i<t&&(o*=256);)n+=this[e+i]*o;return n>=(o*=128)&&(n-=Math.pow(2,8*t)),n},u.prototype.readIntBE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=t,o=1,i=this[e+--n];n>0&&(o*=256);)i+=this[e+--n]*o;return i>=(o*=128)&&(i-=Math.pow(2,8*t)),i},u.prototype.readInt8=function(e,t){return t||D(e,1,this.length),128&this[e]?-1*(255-this[e]+1):this[e]},u.prototype.readInt16LE=function(e,t){t||D(e,2,this.length);var r=this[e]|this[e+1]<<8;return 32768&r?4294901760|r:r},u.prototype.readInt16BE=function(e,t){t||D(e,2,this.length);var r=this[e+1]|this[e]<<8;return 32768&r?4294901760|r:r},u.prototype.readInt32LE=function(e,t){return t||D(e,4,this.length),this[e]|this[e+1]<<8|this[e+2]<<16|this[e+3]<<24},u.prototype.readInt32BE=function(e,t){return t||D(e,4,this.length),this[e]<<24|this[e+1]<<16|this[e+2]<<8|this[e+3]},u.prototype.readFloatLE=function(e,t){return t||D(e,4,this.length),o.read(this,e,!0,23,4)},u.prototype.readFloatBE=function(e,t){return t||D(e,4,this.length),o.read(this,e,!1,23,4)},u.prototype.readDoubleLE=function(e,t){return t||D(e,8,this.length),o.read(this,e,!0,52,8)},u.prototype.readDoubleBE=function(e,t){return t||D(e,8,this.length),o.read(this,e,!1,52,8)},u.prototype.writeUIntLE=function(e,t,r,n){(e=+e,t|=0,r|=0,n)||I(this,e,t,r,Math.pow(2,8*r)-1,0);var o=1,i=0;for(this[t]=255&e;++i<r&&(o*=256);)this[t+i]=e/o&255;return t+r},u.prototype.writeUIntBE=function(e,t,r,n){(e=+e,t|=0,r|=0,n)||I(this,e,t,r,Math.pow(2,8*r)-1,0);var o=r-1,i=1;for(this[t+o]=255&e;--o>=0&&(i*=256);)this[t+o]=e/i&255;return t+r},u.prototype.writeUInt8=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,1,255,0),u.TYPED_ARRAY_SUPPORT||(e=Math.floor(e)),this[t]=255&e,t+1},u.prototype.writeUInt16LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,65535,0),u.TYPED_ARRAY_SUPPORT?(this[t]=255&e,this[t+1]=e>>>8):E(this,e,t,!0),t+2},u.prototype.writeUInt16BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,65535,0),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>8,this[t+1]=255&e):E(this,e,t,!1),t+2},u.prototype.writeUInt32LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,4294967295,0),u.TYPED_ARRAY_SUPPORT?(this[t+3]=e>>>24,this[t+2]=e>>>16,this[t+1]=e>>>8,this[t]=255&e):b(this,e,t,!0),t+4},u.prototype.writeUInt32BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,4294967295,0),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>24,this[t+1]=e>>>16,this[t+2]=e>>>8,this[t+3]=255&e):b(this,e,t,!1),t+4},u.prototype.writeIntLE=function(e,t,r,n){if(e=+e,t|=0,!n){var o=Math.pow(2,8*r-1);I(this,e,t,r,o-1,-o)}var i=0,a=1,s=0;for(this[t]=255&e;++i<r&&(a*=256);)e<0&&0===s&&0!==this[t+i-1]&&(s=1),this[t+i]=(e/a>>0)-s&255;return t+r},u.prototype.writeIntBE=function(e,t,r,n){if(e=+e,t|=0,!n){var o=Math.pow(2,8*r-1);I(this,e,t,r,o-1,-o)}var i=r-1,a=1,s=0;for(this[t+i]=255&e;--i>=0&&(a*=256);)e<0&&0===s&&0!==this[t+i+1]&&(s=1),this[t+i]=(e/a>>0)-s&255;return t+r},u.prototype.writeInt8=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,1,127,-128),u.TYPED_ARRAY_SUPPORT||(e=Math.floor(e)),e<0&&(e=255+e+1),this[t]=255&e,t+1},u.prototype.writeInt16LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,32767,-32768),u.TYPED_ARRAY_SUPPORT?(this[t]=255&e,this[t+1]=e>>>8):E(this,e,t,!0),t+2},u.prototype.writeInt16BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,32767,-32768),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>8,this[t+1]=255&e):E(this,e,t,!1),t+2},u.prototype.writeInt32LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,2147483647,-2147483648),u.TYPED_ARRAY_SUPPORT?(this[t]=255&e,this[t+1]=e>>>8,this[t+2]=e>>>16,this[t+3]=e>>>24):b(this,e,t,!0),t+4},u.prototype.writeInt32BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,2147483647,-2147483648),e<0&&(e=4294967295+e+1),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>24,this[t+1]=e>>>16,this[t+2]=e>>>8,this[t+3]=255&e):b(this,e,t,!1),t+4},u.prototype.writeFloatLE=function(e,t,r){return M(this,e,t,!0,r)},u.prototype.writeFloatBE=function(e,t,r){return M(this,e,t,!1,r)},u.prototype.writeDoubleLE=function(e,t,r){return A(this,e,t,!0,r)},u.prototype.writeDoubleBE=function(e,t,r){return A(this,e,t,!1,r)},u.prototype.copy=function(e,t,r,n){if(r||(r=0),n||0===n||(n=this.length),t>=e.length&&(t=e.length),t||(t=0),n>0&&n<r&&(n=r),n===r)return 0;if(0===e.length||0===this.length)return 0;if(t<0)throw new RangeError("targetStart out of bounds");if(r<0||r>=this.length)throw new RangeError("sourceStart out of bounds");if(n<0)throw new RangeError("sourceEnd out of bounds");n>this.length&&(n=this.length),e.length-t<n-r&&(n=e.length-t+r);var o,i=n-r;if(this===e&&r<t&&t<n)for(o=i-1;o>=0;--o)e[o+t]=this[o+r];else if(i<1e3||!u.TYPED_ARRAY_SUPPORT)for(o=0;o<i;++o)e[o+t]=this[o+r];else Uint8Array.prototype.set.call(e,this.subarray(r,r+i),t);return i},u.prototype.fill=function(e,t,r,n){if("string"==typeof e){if("string"==typeof t?(n=t,t=0,r=this.length):"string"==typeof r&&(n=r,r=this.length),1===e.length){var o=e.charCodeAt(0);o<256&&(e=o)}if(void 0!==n&&"string"!=typeof n)throw new TypeError("encoding must be a string");if("string"==typeof n&&!u.isEncoding(n))throw new TypeError("Unknown encoding: "+n)}else"number"==typeof e&&(e&=255);if(t<0||this.length<t||this.length<r)throw new RangeError("Out of range index");if(r<=t)return this;var i;if(t>>>=0,r=void 0===r?this.length:r>>>0,e||(e=0),"number"==typeof e)for(i=t;i<r;++i)this[i]=e;else{var a=u.isBuffer(e)?e:O(new u(e,n).toString()),s=a.length;for(i=0;i<r-t;++i)this[i+t]=a[i%s]}return this};var U=/[^+\/0-9A-Za-z-_]/g;function B(e){return e<16?"0"+e.toString(16):e.toString(16)}function O(e,t){var r;t=t||1/0;for(var n=e.length,o=null,i=[],a=0;a<n;++a){if((r=e.charCodeAt(a))>55295&&r<57344){if(!o){if(r>56319){(t-=3)>-1&&i.push(239,191,189);continue}if(a+1===n){(t-=3)>-1&&i.push(239,191,189);continue}o=r;continue}if(r<56320){(t-=3)>-1&&i.push(239,191,189),o=r;continue}r=65536+(o-55296<<10|r-56320)}else o&&(t-=3)>-1&&i.push(239,191,189);if(o=null,r<128){if((t-=1)<0)break;i.push(r)}else if(r<2048){if((t-=2)<0)break;i.push(r>>6|192,63&r|128)}else if(r<65536){if((t-=3)<0)break;i.push(r>>12|224,r>>6&63|128,63&r|128)}else{if(!(r<1114112))throw new Error("Invalid code point");if((t-=4)<0)break;i.push(r>>18|240,r>>12&63|128,r>>6&63|128,63&r|128)}}return i}function L(e){return n.toByteArray(function(e){var t;if((e=(t=e,t.trim?t.trim():t.replace(/^\s+|\s+$/g,"")).replace(U,"")).length<2)return"";for(;e.length%4!=0;)e+="=";return e}(e))}function q(e,t,r,n){for(var o=0;o<n&&!(o+r>=t.length||o>=e.length);++o)t[o+r]=e[o];return o}}).call(t,r(2))},function(e,t){var r;r=function(){return this}();try{r=r||Function("return this")()||(0,eval)("this")}catch(e){"object"==typeof window&&(r=window)}e.exports=r},function(e,t,r){"use strict";t.byteLength=function(e){return 3*e.length/4-c(e)},t.toByteArray=function(e){var t,r,n,a,s,u=e.length;a=c(e),s=new i(3*u/4-a),r=a>0?u-4:u;var l=0;for(t=0;t<r;t+=4)n=o[e.charCodeAt(t)]<<18|o[e.charCodeAt(t+1)]<<12|o[e.charCodeAt(t+2)]<<6|o[e.charCodeAt(t+3)],s[l++]=n>>16&255,s[l++]=n>>8&255,s[l++]=255&n;2===a?(n=o[e.charCodeAt(t)]<<2|o[e.charCodeAt(t+1)]>>4,s[l++]=255&n):1===a&&(n=o[e.charCodeAt(t)]<<10|o[e.charCodeAt(t+1)]<<4|o[e.charCodeAt(t+2)]>>2,s[l++]=n>>8&255,s[l++]=255&n);return s},t.fromByteArray=function(e){for(var t,r=e.length,o=r%3,i="",a=[],s=0,u=r-o;s<u;s+=16383)a.push(l(e,s,s+16383>u?u:s+16383));1===o?(t=e[r-1],i+=n[t>>2],i+=n[t<<4&63],i+="=="):2===o&&(t=(e[r-2]<<8)+e[r-1],i+=n[t>>10],i+=n[t>>4&63],i+=n[t<<2&63],i+="=");return a.push(i),a.join("")};for(var n=[],o=[],i="undefined"!=typeof Uint8Array?Uint8Array:Array,a="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",s=0,u=a.length;s<u;++s)n[s]=a[s],o[a.charCodeAt(s)]=s;function c(e){var t=e.length;if(t%4>0)throw new Error("Invalid string. Length must be a multiple of 4");return"="===e[t-2]?2:"="===e[t-1]?1:0}function l(e,t,r){for(var o,i,a=[],s=t;s<r;s+=3)o=(e[s]<<16)+(e[s+1]<<8)+e[s+2],a.push(n[(i=o)>>18&63]+n[i>>12&63]+n[i>>6&63]+n[63&i]);return a.join("")}o["-".charCodeAt(0)]=62,o["_".charCodeAt(0)]=63},function(e,t){t.read=function(e,t,r,n,o){var i,a,s=8*o-n-1,u=(1<<s)-1,c=u>>1,l=-7,m=r?o-1:0,f=r?-1:1,p=e[t+m];for(m+=f,i=p&(1<<-l)-1,p>>=-l,l+=s;l>0;i=256*i+e[t+m],m+=f,l-=8);for(a=i&(1<<-l)-1,i>>=-l,l+=n;l>0;a=256*a+e[t+m],m+=f,l-=8);if(0===i)i=1-c;else{if(i===u)return a?NaN:1/0*(p?-1:1);a+=Math.pow(2,n),i-=c}return(p?-1:1)*a*Math.pow(2,i-n)},t.write=function(e,t,r,n,o,i){var a,s,u,c=8*i-o-1,l=(1<<c)-1,m=l>>1,f=23===o?Math.pow(2,-24)-Math.pow(2,-77):0,p=n?0:i-1,h=n?1:-1,y=t<0||0===t&&1/t<0?1:0;for(t=Math.abs(t),isNaN(t)||t===1/0?(s=isNaN(t)?1:0,a=l):(a=Math.floor(Math.log(t)/Math.LN2),t*(u=Math.pow(2,-a))<1&&(a--,u*=2),(t+=a+m>=1?f/u:f*Math.pow(2,1-m))*u>=2&&(a++,u/=2),a+m>=l?(s=0,a=l):a+m>=1?(s=(t*u-1)*Math.pow(2,o),a+=m):(s=t*Math.pow(2,m-1)*Math.pow(2,o),a=0));o>=8;e[r+p]=255&s,p+=h,s/=256,o-=8);for(a=a<<o|s,c+=o;c>0;e[r+p]=255&a,p+=h,a/=256,c-=8);e[r+p-h]|=128*y}},function(e,t){var r={}.toString;e.exports=Array.isArray||function(e){return"[object Array]"==r.call(e)}},function(t,r){t.exports=e},function(e,r){e.exports=t},function(e,t,r){"use strict";e.exports={Alert:{name:"Alert"},AuthLDAP:{name:"AuthLDAP"},Computer:{name:"Computer"},Config:{name:"Config"},ConsumableItem:{name:"ConsumableItem"},Contact:{name:"Contact"},Contract:{name:"Contract"},CronTask:{name:"CronTask"},CronTaskLog:{name:"CronTaskLog"},DBConnection:{name:"DBConnection"},DisplayPreference:{name:"DisplayPreference"},Document:{name:"Document"},AuthLdapReplicate:{name:"AuthLdapReplicate"},Event:{name:"Event"},KnowbaseItem:{name:"KnowbaseItem"},Link:{name:"Link"},Log:{name:"Log"},MailCollector:{name:"MailCollector"},Monitor:{name:"Monitor"},NetworkEquipment:{name:"NetworkEquipment"},Notification:{name:"Notification"},NotificationEvent:{name:"NotificationEvent"},NotificationMailSetting:{name:"NotificationMailSetting"},AuthMail:{name:"AuthMail"},NotificationTemplate:{name:"NotificationTemplate"},NotImportedEmail:{name:"NotImportedEmail"},Peripheral:{name:"Peripheral"},Phone:{name:"Phone"},Plugin:{name:"Plugin"},Printer:{name:"Printer"},Profile:{name:"Profile"},Project:{name:"Project"},QueuedMail:{name:"QueuedMail"},Reminder:{name:"Reminder"},Bookmark:{name:"Bookmark"},RSSFeed:{name:"RSSFeed"},Rule:{name:"Rule"},RuleCollection:{name:"RuleCollection"},SLA:{name:"SLA"},SlaLevel_Ticket:{name:"SlaLevel_Ticket"},Software:{name:"Software"},SoftwareLicense:{name:"SoftwareLicense"},Supplier:{name:"Supplier"},TicketFollowup:{name:"TicketFollowup"},TicketSatisfaction:{name:"TicketSatisfaction"},CartridgeItem:{name:"CartridgeItem"},Transfer:{name:"Transfer"},User:{name:"User"},CommonDBConnexity:{name:"CommonDBConnexity"},CommonDropdown:{name:"CommonDropdown"},CommonITILObject:{name:"CommonITILObject"},CommonITILTask:{name:"CommonITILTask"},AutoUpdateSystem:{name:"AutoUpdateSystem"},Blacklist:{name:"Blacklist"},CartridgeItemType:{name:"CartridgeItemType"},Item_DeviceSoundCard:{name:"Item_DeviceSoundCard"},Item_Problem:{name:"Item_Problem"},Item_Project:{name:"Item_Project"},Item_Ticket:{name:"Item_Ticket"},ITILCategory:{name:"ITILCategory"},KnowbaseItem_Profile:{name:"KnowbaseItem_Profile"},KnowbaseItem_User:{name:"KnowbaseItem_User"},KnowbaseItemCategory:{name:"KnowbaseItemCategory"},KnowbaseItemTranslation:{name:"KnowbaseItemTranslation"},Link_Itemtype:{name:"Link_Itemtype"},Change:{name:"Change"},Location:{name:"Location"},Manufacturer:{name:"Manufacturer"},MonitorModel:{name:"MonitorModel"},MonitorType:{name:"MonitorType"},Netpoint:{name:"Netpoint"},Network:{name:"Network"},NetworkAlias:{name:"NetworkAlias"},NetworkEquipmentFirmware:{name:"NetworkEquipmentFirmware"},NetworkEquipmentModel:{name:"NetworkEquipmentModel"},NetworkEquipmentType:{name:"NetworkEquipmentType"},Change_Group:{name:"Change_Group"},NetworkInterface:{name:"NetworkInterface"},NetworkName:{name:"NetworkName"},NetworkPort:{name:"NetworkPort"},NetworkPort_NetworkPort:{name:"NetworkPort_NetworkPort"},NetworkPort_Vlan:{name:"NetworkPort_Vlan"},NetworkPortAggregate:{name:"NetworkPortAggregate"},NetworkPortAlias:{name:"NetworkPortAlias"},NetworkPortDialup:{name:"NetworkPortDialup"},NetworkPortEthernet:{name:"NetworkPortEthernet"},NetworkPortInstantiation:{name:"NetworkPortInstantiation"},Change_Item:{name:"Change_Item"},NetworkPortLocal:{name:"NetworkPortLocal"},NetworkPortMigration:{name:"NetworkPortMigration"},NetworkPortWifi:{name:"NetworkPortWifi"},Notepad:{name:"Notepad"},NotificationTarget:{name:"NotificationTarget"},NotificationTargetCartridgeItem:{name:"NotificationTargetCartridgeItem"},NotificationTargetChange:{name:"NotificationTargetChange"},NotificationTargetCommonITILObject:{name:"NotificationTargetCommonITILObject"},NotificationTargetConsumableItem:{name:"NotificationTargetConsumableItem"},NotificationTargetContract:{name:"NotificationTargetContract"},Change_Problem:{name:"Change_Problem"},NotificationTargetCrontask:{name:"NotificationTargetCrontask"},NotificationTargetDBConnection:{name:"NotificationTargetDBConnection"},NotificationTargetFieldUnicity:{name:"NotificationTargetFieldUnicity"},NotificationTargetInfocom:{name:"NotificationTargetInfocom"},NotificationTargetMailCollector:{name:"NotificationTargetMailCollector"},NotificationTargetPlanningRecall:{name:"NotificationTargetPlanningRecall"},NotificationTargetProblem:{name:"NotificationTargetProblem"},NotificationTargetProject:{name:"NotificationTargetProject"},NotificationTargetProjectTask:{name:"NotificationTargetProjectTask"},NotificationTargetReservation:{name:"NotificationTargetReservation"},Change_Project:{name:"Change_Project"},NotificationTargetSoftwareLicense:{name:"NotificationTargetSoftwareLicense"},NotificationTargetTicket:{name:"NotificationTargetTicket"},NotificationTargetUser:{name:"NotificationTargetUser"},NotificationTemplateTranslation:{name:"NotificationTemplateTranslation"},OperatingSystem:{name:"OperatingSystem"},OperatingSystemServicePack:{name:"OperatingSystemServicePack"},OperatingSystemVersion:{name:"OperatingSystemVersion"},PeripheralModel:{name:"PeripheralModel"},PeripheralType:{name:"PeripheralType"},PhoneModel:{name:"PhoneModel"},Change_Supplier:{name:"Change_Supplier"},PhonePowerSupply:{name:"PhonePowerSupply"},PhoneType:{name:"PhoneType"},PlanningRecall:{name:"PlanningRecall"},PrinterModel:{name:"PrinterModel"},PrinterType:{name:"PrinterType"},Problem:{name:"Problem"},Problem_Supplier:{name:"Problem_Supplier"},Problem_Ticket:{name:"Problem_Ticket"},Problem_User:{name:"Problem_User"},ProblemCost:{name:"ProblemCost"},Change_Ticket:{name:"Change_Ticket"},ProblemTask:{name:"ProblemTask"},Profile_Reminder:{name:"Profile_Reminder"},Profile_RSSFeed:{name:"Profile_RSSFeed"},Profile_User:{name:"Profile_User"},ProfileRight:{name:"ProfileRight"},ProjectCost:{name:"ProjectCost"},ProjectState:{name:"ProjectState"},ProjectTask:{name:"ProjectTask"},ProjectTask_Ticket:{name:"ProjectTask_Ticket"},ProjectTaskTeam:{name:"ProjectTaskTeam"},Change_User:{name:"Change_User"},ProjectTaskType:{name:"ProjectTaskType"},ProjectTeam:{name:"ProjectTeam"},ProjectType:{name:"ProjectType"},RegisteredID:{name:"RegisteredID"},Reminder_User:{name:"Reminder_User"},RequestType:{name:"RequestType"},Reservation:{name:"Reservation"},ReservationItem:{name:"ReservationItem"},RSSFeed_User:{name:"RSSFeed_User"},RuleAction:{name:"RuleAction"},ChangeCost:{name:"ChangeCost"},RuleCriteria:{name:"RuleCriteria"},RuleDictionnaryComputerModel:{name:"RuleDictionnaryComputerModel"},RuleDictionnaryComputerModelCollection:{name:"RuleDictionnaryComputerModelCollection"},RuleDictionnaryComputerType:{name:"RuleDictionnaryComputerType"},RuleDictionnaryComputerTypeCollection:{name:"RuleDictionnaryComputerTypeCollection"},RuleDictionnaryDropdown:{name:"RuleDictionnaryDropdown"},RuleDictionnaryDropdownCollection:{name:"RuleDictionnaryDropdownCollection"},RuleDictionnaryManufacturer:{name:"RuleDictionnaryManufacturer"},RuleDictionnaryManufacturerCollection:{name:"RuleDictionnaryManufacturerCollection"},RuleDictionnaryMonitorModel:{name:"RuleDictionnaryMonitorModel"},BlacklistedMailContent:{name:"BlacklistedMailContent"},ChangeTask:{name:"ChangeTask"},RuleDictionnaryMonitorModelCollection:{name:"RuleDictionnaryMonitorModelCollection"},RuleDictionnaryMonitorType:{name:"RuleDictionnaryMonitorType"},RuleDictionnaryMonitorTypeCollection:{name:"RuleDictionnaryMonitorTypeCollection"},RuleDictionnaryNetworkEquipmentModel:{name:"RuleDictionnaryNetworkEquipmentModel"},RuleDictionnaryNetworkEquipmentModelCollection:{name:"RuleDictionnaryNetworkEquipmentModelCollection"},RuleDictionnaryNetworkEquipmentType:{name:"RuleDictionnaryNetworkEquipmentType"},RuleDictionnaryNetworkEquipmentTypeCollection:{name:"RuleDictionnaryNetworkEquipmentTypeCollection"},RuleDictionnaryOperatingSystem:{name:"RuleDictionnaryOperatingSystem"},RuleDictionnaryOperatingSystemCollection:{name:"RuleDictionnaryOperatingSystemCollection"},RuleDictionnaryOperatingSystemServicePack:{name:"RuleDictionnaryOperatingSystemServicePack"},ChangeValidation:{name:"ChangeValidation"},RuleDictionnaryOperatingSystemServicePackCollection:{name:"RuleDictionnaryOperatingSystemServicePackCollection"},RuleDictionnaryOperatingSystemVersion:{name:"RuleDictionnaryOperatingSystemVersion"},RuleDictionnaryOperatingSystemVersionCollection:{name:"RuleDictionnaryOperatingSystemVersionCollection"},RuleDictionnaryPeripheralModel:{name:"RuleDictionnaryPeripheralModel"},RuleDictionnaryPeripheralModelCollection:{name:"RuleDictionnaryPeripheralModelCollection"},RuleDictionnaryPeripheralType:{name:"RuleDictionnaryPeripheralType"},RuleDictionnaryPeripheralTypeCollection:{name:"RuleDictionnaryPeripheralTypeCollection"},RuleDictionnaryPhoneModel:{name:"RuleDictionnaryPhoneModel"},RuleDictionnaryPhoneModelCollection:{name:"RuleDictionnaryPhoneModelCollection"},RuleDictionnaryPhoneType:{name:"RuleDictionnaryPhoneType"},CommonDBChild:{name:"CommonDBChild"},RuleDictionnaryPhoneTypeCollection:{name:"RuleDictionnaryPhoneTypeCollection"},RuleDictionnaryPrinter:{name:"RuleDictionnaryPrinter"},RuleDictionnaryPrinterCollection:{name:"RuleDictionnaryPrinterCollection"},RuleDictionnaryPrinterModel:{name:"RuleDictionnaryPrinterModel"},RuleDictionnaryPrinterModelCollection:{name:"RuleDictionnaryPrinterModelCollection"},RuleDictionnaryPrinterType:{name:"RuleDictionnaryPrinterType"},RuleDictionnaryPrinterTypeCollection:{name:"RuleDictionnaryPrinterTypeCollection"},RuleDictionnarySoftware:{name:"RuleDictionnarySoftware"},RuleDictionnarySoftwareCollection:{name:"RuleDictionnarySoftwareCollection"},RuleImportComputer:{name:"RuleImportComputer"},CommonDBRelation:{name:"CommonDBRelation"},RuleImportComputerCollection:{name:"RuleImportComputerCollection"},RuleImportEntity:{name:"RuleImportEntity"},RuleImportEntityCollection:{name:"RuleImportEntityCollection"},RuleMailCollector:{name:"RuleMailCollector"},RuleMailCollectorCollection:{name:"RuleMailCollectorCollection"},RuleRight:{name:"RuleRight"},RuleRightCollection:{name:"RuleRightCollection"},RuleRightParameter:{name:"RuleRightParameter"},RuleSoftwareCategory:{name:"RuleSoftwareCategory"},RuleSoftwareCategoryCollection:{name:"RuleSoftwareCategoryCollection"},CommonDevice:{name:"CommonDevice"},RuleTicket:{name:"RuleTicket"},RuleTicketCollection:{name:"RuleTicketCollection"},SlaLevel:{name:"SlaLevel"},SlaLevelAction:{name:"SlaLevelAction"},SlaLevelCriteria:{name:"SlaLevelCriteria"},SoftwareCategory:{name:"SoftwareCategory"},SoftwareLicenseType:{name:"SoftwareLicenseType"},SoftwareVersion:{name:"SoftwareVersion"},SolutionTemplate:{name:"SolutionTemplate"},SolutionType:{name:"SolutionType"},CommonImplicitTreeDropdown:{name:"CommonImplicitTreeDropdown"},SsoVariable:{name:"SsoVariable"},State:{name:"State"},Supplier_Ticket:{name:"Supplier_Ticket"},SupplierType:{name:"SupplierType"},TaskCategory:{name:"TaskCategory"},Ticket:{name:"Ticket"},Ticket_Ticket:{name:"Ticket_Ticket"},Ticket_User:{name:"Ticket_User"},TicketCost:{name:"TicketCost"},TicketRecurrent:{name:"TicketRecurrent"},CommonITILActor:{name:"CommonITILActor"},TicketTask:{name:"TicketTask"},TicketTemplate:{name:"TicketTemplate"},TicketTemplateHiddenField:{name:"TicketTemplateHiddenField"},TicketTemplateMandatoryField:{name:"TicketTemplateMandatoryField"},TicketTemplatePredefinedField:{name:"TicketTemplatePredefinedField"},TicketValidation:{name:"TicketValidation"},UserCategory:{name:"UserCategory"},UserEmail:{name:"UserEmail"},UserTitle:{name:"UserTitle"},VirtualMachineState:{name:"VirtualMachineState"},CommonITILCost:{name:"CommonITILCost"},VirtualMachineSystem:{name:"VirtualMachineSystem"},VirtualMachineType:{name:"VirtualMachineType"},Vlan:{name:"Vlan"},WifiNetwork:{name:"WifiNetwork"},CommonITILValidation:{name:"CommonITILValidation"},CommonTreeDropdown:{name:"CommonTreeDropdown"},Bookmark_User:{name:"Bookmark_User"},Computer_Item:{name:"Computer_Item"},Computer_SoftwareLicense:{name:"Computer_SoftwareLicense"},Computer_SoftwareVersion:{name:"Computer_SoftwareVersion"},ComputerDisk:{name:"ComputerDisk"},ComputerModel:{name:"ComputerModel"},ComputerType:{name:"ComputerType"},ComputerVirtualMachine:{name:"ComputerVirtualMachine"},Consumable:{name:"Consumable"},ConsumableItemType:{name:"ConsumableItemType"},Contact_Supplier:{name:"Contact_Supplier"},Budget:{name:"Budget"},ContactType:{name:"ContactType"},Contract_Item:{name:"Contract_Item"},Contract_Supplier:{name:"Contract_Supplier"},ContractCost:{name:"ContractCost"},ContractType:{name:"ContractType"},DeviceCase:{name:"DeviceCase"},DeviceCaseType:{name:"DeviceCaseType"},DeviceControl:{name:"DeviceControl"},DeviceDrive:{name:"DeviceDrive"},DeviceGraphicCard:{name:"DeviceGraphicCard"},Calendar:{name:"Calendar"},DeviceHardDrive:{name:"DeviceHardDrive"},DeviceMemory:{name:"DeviceMemory"},DeviceMemoryType:{name:"DeviceMemoryType"},DeviceMotherboard:{name:"DeviceMotherboard"},DeviceNetworkCard:{name:"DeviceNetworkCard"},DevicePci:{name:"DevicePci"},DevicePowerSupply:{name:"DevicePowerSupply"},DeviceProcessor:{name:"DeviceProcessor"},DeviceSoundCard:{name:"DeviceSoundCard"},Document_Item:{name:"Document_Item"},Calendar_Holiday:{name:"Calendar_Holiday"},DocumentCategory:{name:"DocumentCategory"},DocumentType:{name:"DocumentType"},Domain:{name:"Domain"},DropdownTranslation:{name:"DropdownTranslation"},Entity:{name:"Entity"},Entity_KnowbaseItem:{name:"Entity_KnowbaseItem"},Entity_Reminder:{name:"Entity_Reminder"},Entity_RSSFeed:{name:"Entity_RSSFeed"},Fieldblacklist:{name:"Fieldblacklist"},FieldUnicity:{name:"FieldUnicity"},CalendarSegment:{name:"CalendarSegment"},Filesystem:{name:"Filesystem"},FQDN:{name:"FQDN"},FQDNLabel:{name:"FQDNLabel"},Group:{name:"Group"},Group_KnowbaseItem:{name:"Group_KnowbaseItem"},Group_Problem:{name:"Group_Problem"},Group_Reminder:{name:"Group_Reminder"},Group_RSSFeed:{name:"Group_RSSFeed"},Group_Ticket:{name:"Group_Ticket"},Group_User:{name:"Group_User"},Cartridge:{name:"Cartridge"},Holiday:{name:"Holiday"},Infocom:{name:"Infocom"},InterfaceType:{name:"InterfaceType"},IPAddress:{name:"IPAddress"},IPAddress_IPNetwork:{name:"IPAddress_IPNetwork"},IPNetmask:{name:"IPNetmask"},IPNetwork:{name:"IPNetwork"},IPNetwork_Vlan:{name:"IPNetwork_Vlan"},Item_DeviceCase:{name:"Item_DeviceCase"},Item_DeviceControl:{name:"Item_DeviceControl"},CartridgeItem_PrinterModel:{name:"CartridgeItem_PrinterModel"},Item_DeviceDrive:{name:"Item_DeviceDrive"},Item_DeviceGraphicCard:{name:"Item_DeviceGraphicCard"},Item_DeviceHardDrive:{name:"Item_DeviceHardDrive"},Item_DeviceMemory:{name:"Item_DeviceMemory"},Item_DeviceMotherboard:{name:"Item_DeviceMotherboard"},Item_DeviceNetworkCard:{name:"Item_DeviceNetworkCard"},Item_DevicePci:{name:"Item_DevicePci"},Item_DevicePowerSupply:{name:"Item_DevicePowerSupply"},Item_DeviceProcessor:{name:"Item_DeviceProcessor"},Item_Devices:{name:"Item_Devices"}}}])});
+!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t(require("https"),require("url")):"function"==typeof define&&define.amd?define(["https","url"],t):"object"==typeof exports?exports.GlpiRestClient=t(require("https"),require("url")):e.GlpiRestClient=t(e.https,e.url)}("undefined"!=typeof self?self:this,function(e,t){return function(e){var t={};function r(n){if(t[n])return t[n].exports;var o=t[n]={i:n,l:!1,exports:{}};return e[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}return r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{configurable:!1,enumerable:!0,get:n})},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=1)}([function(e,t,r){"use strict";e.exports={Alert:{name:"Alert"},AuthLDAP:{name:"AuthLDAP"},Computer:{name:"Computer"},Config:{name:"Config"},ConsumableItem:{name:"ConsumableItem"},Contact:{name:"Contact"},Contract:{name:"Contract"},CronTask:{name:"CronTask"},CronTaskLog:{name:"CronTaskLog"},DBConnection:{name:"DBConnection"},DisplayPreference:{name:"DisplayPreference"},Document:{name:"Document"},AuthLdapReplicate:{name:"AuthLdapReplicate"},Event:{name:"Event"},KnowbaseItem:{name:"KnowbaseItem"},Link:{name:"Link"},Log:{name:"Log"},MailCollector:{name:"MailCollector"},Monitor:{name:"Monitor"},NetworkEquipment:{name:"NetworkEquipment"},Notification:{name:"Notification"},NotificationEvent:{name:"NotificationEvent"},NotificationMailSetting:{name:"NotificationMailSetting"},AuthMail:{name:"AuthMail"},NotificationTemplate:{name:"NotificationTemplate"},NotImportedEmail:{name:"NotImportedEmail"},Peripheral:{name:"Peripheral"},Phone:{name:"Phone"},Plugin:{name:"Plugin"},Printer:{name:"Printer"},Profile:{name:"Profile"},Project:{name:"Project"},QueuedMail:{name:"QueuedMail"},Reminder:{name:"Reminder"},Bookmark:{name:"Bookmark"},RSSFeed:{name:"RSSFeed"},Rule:{name:"Rule"},RuleCollection:{name:"RuleCollection"},SLA:{name:"SLA"},SlaLevel_Ticket:{name:"SlaLevel_Ticket"},Software:{name:"Software"},SoftwareLicense:{name:"SoftwareLicense"},Supplier:{name:"Supplier"},TicketFollowup:{name:"TicketFollowup"},TicketSatisfaction:{name:"TicketSatisfaction"},CartridgeItem:{name:"CartridgeItem"},Transfer:{name:"Transfer"},User:{name:"User"},CommonDBConnexity:{name:"CommonDBConnexity"},CommonDropdown:{name:"CommonDropdown"},CommonITILObject:{name:"CommonITILObject"},CommonITILTask:{name:"CommonITILTask"},AutoUpdateSystem:{name:"AutoUpdateSystem"},Blacklist:{name:"Blacklist"},CartridgeItemType:{name:"CartridgeItemType"},Item_DeviceSoundCard:{name:"Item_DeviceSoundCard"},Item_Problem:{name:"Item_Problem"},Item_Project:{name:"Item_Project"},Item_Ticket:{name:"Item_Ticket"},ITILCategory:{name:"ITILCategory"},KnowbaseItem_Profile:{name:"KnowbaseItem_Profile"},KnowbaseItem_User:{name:"KnowbaseItem_User"},KnowbaseItemCategory:{name:"KnowbaseItemCategory"},KnowbaseItemTranslation:{name:"KnowbaseItemTranslation"},Link_Itemtype:{name:"Link_Itemtype"},Change:{name:"Change"},Location:{name:"Location"},Manufacturer:{name:"Manufacturer"},MonitorModel:{name:"MonitorModel"},MonitorType:{name:"MonitorType"},Netpoint:{name:"Netpoint"},Network:{name:"Network"},NetworkAlias:{name:"NetworkAlias"},NetworkEquipmentFirmware:{name:"NetworkEquipmentFirmware"},NetworkEquipmentModel:{name:"NetworkEquipmentModel"},NetworkEquipmentType:{name:"NetworkEquipmentType"},Change_Group:{name:"Change_Group"},NetworkInterface:{name:"NetworkInterface"},NetworkName:{name:"NetworkName"},NetworkPort:{name:"NetworkPort"},NetworkPort_NetworkPort:{name:"NetworkPort_NetworkPort"},NetworkPort_Vlan:{name:"NetworkPort_Vlan"},NetworkPortAggregate:{name:"NetworkPortAggregate"},NetworkPortAlias:{name:"NetworkPortAlias"},NetworkPortDialup:{name:"NetworkPortDialup"},NetworkPortEthernet:{name:"NetworkPortEthernet"},NetworkPortInstantiation:{name:"NetworkPortInstantiation"},Change_Item:{name:"Change_Item"},NetworkPortLocal:{name:"NetworkPortLocal"},NetworkPortMigration:{name:"NetworkPortMigration"},NetworkPortWifi:{name:"NetworkPortWifi"},Notepad:{name:"Notepad"},NotificationTarget:{name:"NotificationTarget"},NotificationTargetCartridgeItem:{name:"NotificationTargetCartridgeItem"},NotificationTargetChange:{name:"NotificationTargetChange"},NotificationTargetCommonITILObject:{name:"NotificationTargetCommonITILObject"},NotificationTargetConsumableItem:{name:"NotificationTargetConsumableItem"},NotificationTargetContract:{name:"NotificationTargetContract"},Change_Problem:{name:"Change_Problem"},NotificationTargetCrontask:{name:"NotificationTargetCrontask"},NotificationTargetDBConnection:{name:"NotificationTargetDBConnection"},NotificationTargetFieldUnicity:{name:"NotificationTargetFieldUnicity"},NotificationTargetInfocom:{name:"NotificationTargetInfocom"},NotificationTargetMailCollector:{name:"NotificationTargetMailCollector"},NotificationTargetPlanningRecall:{name:"NotificationTargetPlanningRecall"},NotificationTargetProblem:{name:"NotificationTargetProblem"},NotificationTargetProject:{name:"NotificationTargetProject"},NotificationTargetProjectTask:{name:"NotificationTargetProjectTask"},NotificationTargetReservation:{name:"NotificationTargetReservation"},Change_Project:{name:"Change_Project"},NotificationTargetSoftwareLicense:{name:"NotificationTargetSoftwareLicense"},NotificationTargetTicket:{name:"NotificationTargetTicket"},NotificationTargetUser:{name:"NotificationTargetUser"},NotificationTemplateTranslation:{name:"NotificationTemplateTranslation"},OperatingSystem:{name:"OperatingSystem"},OperatingSystemServicePack:{name:"OperatingSystemServicePack"},OperatingSystemVersion:{name:"OperatingSystemVersion"},PeripheralModel:{name:"PeripheralModel"},PeripheralType:{name:"PeripheralType"},PhoneModel:{name:"PhoneModel"},Change_Supplier:{name:"Change_Supplier"},PhonePowerSupply:{name:"PhonePowerSupply"},PhoneType:{name:"PhoneType"},PlanningRecall:{name:"PlanningRecall"},PrinterModel:{name:"PrinterModel"},PrinterType:{name:"PrinterType"},Problem:{name:"Problem"},Problem_Supplier:{name:"Problem_Supplier"},Problem_Ticket:{name:"Problem_Ticket"},Problem_User:{name:"Problem_User"},ProblemCost:{name:"ProblemCost"},Change_Ticket:{name:"Change_Ticket"},ProblemTask:{name:"ProblemTask"},Profile_Reminder:{name:"Profile_Reminder"},Profile_RSSFeed:{name:"Profile_RSSFeed"},Profile_User:{name:"Profile_User"},ProfileRight:{name:"ProfileRight"},ProjectCost:{name:"ProjectCost"},ProjectState:{name:"ProjectState"},ProjectTask:{name:"ProjectTask"},ProjectTask_Ticket:{name:"ProjectTask_Ticket"},ProjectTaskTeam:{name:"ProjectTaskTeam"},Change_User:{name:"Change_User"},ProjectTaskType:{name:"ProjectTaskType"},ProjectTeam:{name:"ProjectTeam"},ProjectType:{name:"ProjectType"},RegisteredID:{name:"RegisteredID"},Reminder_User:{name:"Reminder_User"},RequestType:{name:"RequestType"},Reservation:{name:"Reservation"},ReservationItem:{name:"ReservationItem"},RSSFeed_User:{name:"RSSFeed_User"},RuleAction:{name:"RuleAction"},ChangeCost:{name:"ChangeCost"},RuleCriteria:{name:"RuleCriteria"},RuleDictionnaryComputerModel:{name:"RuleDictionnaryComputerModel"},RuleDictionnaryComputerModelCollection:{name:"RuleDictionnaryComputerModelCollection"},RuleDictionnaryComputerType:{name:"RuleDictionnaryComputerType"},RuleDictionnaryComputerTypeCollection:{name:"RuleDictionnaryComputerTypeCollection"},RuleDictionnaryDropdown:{name:"RuleDictionnaryDropdown"},RuleDictionnaryDropdownCollection:{name:"RuleDictionnaryDropdownCollection"},RuleDictionnaryManufacturer:{name:"RuleDictionnaryManufacturer"},RuleDictionnaryManufacturerCollection:{name:"RuleDictionnaryManufacturerCollection"},RuleDictionnaryMonitorModel:{name:"RuleDictionnaryMonitorModel"},BlacklistedMailContent:{name:"BlacklistedMailContent"},ChangeTask:{name:"ChangeTask"},RuleDictionnaryMonitorModelCollection:{name:"RuleDictionnaryMonitorModelCollection"},RuleDictionnaryMonitorType:{name:"RuleDictionnaryMonitorType"},RuleDictionnaryMonitorTypeCollection:{name:"RuleDictionnaryMonitorTypeCollection"},RuleDictionnaryNetworkEquipmentModel:{name:"RuleDictionnaryNetworkEquipmentModel"},RuleDictionnaryNetworkEquipmentModelCollection:{name:"RuleDictionnaryNetworkEquipmentModelCollection"},RuleDictionnaryNetworkEquipmentType:{name:"RuleDictionnaryNetworkEquipmentType"},RuleDictionnaryNetworkEquipmentTypeCollection:{name:"RuleDictionnaryNetworkEquipmentTypeCollection"},RuleDictionnaryOperatingSystem:{name:"RuleDictionnaryOperatingSystem"},RuleDictionnaryOperatingSystemCollection:{name:"RuleDictionnaryOperatingSystemCollection"},RuleDictionnaryOperatingSystemServicePack:{name:"RuleDictionnaryOperatingSystemServicePack"},ChangeValidation:{name:"ChangeValidation"},RuleDictionnaryOperatingSystemServicePackCollection:{name:"RuleDictionnaryOperatingSystemServicePackCollection"},RuleDictionnaryOperatingSystemVersion:{name:"RuleDictionnaryOperatingSystemVersion"},RuleDictionnaryOperatingSystemVersionCollection:{name:"RuleDictionnaryOperatingSystemVersionCollection"},RuleDictionnaryPeripheralModel:{name:"RuleDictionnaryPeripheralModel"},RuleDictionnaryPeripheralModelCollection:{name:"RuleDictionnaryPeripheralModelCollection"},RuleDictionnaryPeripheralType:{name:"RuleDictionnaryPeripheralType"},RuleDictionnaryPeripheralTypeCollection:{name:"RuleDictionnaryPeripheralTypeCollection"},RuleDictionnaryPhoneModel:{name:"RuleDictionnaryPhoneModel"},RuleDictionnaryPhoneModelCollection:{name:"RuleDictionnaryPhoneModelCollection"},RuleDictionnaryPhoneType:{name:"RuleDictionnaryPhoneType"},CommonDBChild:{name:"CommonDBChild"},RuleDictionnaryPhoneTypeCollection:{name:"RuleDictionnaryPhoneTypeCollection"},RuleDictionnaryPrinter:{name:"RuleDictionnaryPrinter"},RuleDictionnaryPrinterCollection:{name:"RuleDictionnaryPrinterCollection"},RuleDictionnaryPrinterModel:{name:"RuleDictionnaryPrinterModel"},RuleDictionnaryPrinterModelCollection:{name:"RuleDictionnaryPrinterModelCollection"},RuleDictionnaryPrinterType:{name:"RuleDictionnaryPrinterType"},RuleDictionnaryPrinterTypeCollection:{name:"RuleDictionnaryPrinterTypeCollection"},RuleDictionnarySoftware:{name:"RuleDictionnarySoftware"},RuleDictionnarySoftwareCollection:{name:"RuleDictionnarySoftwareCollection"},RuleImportComputer:{name:"RuleImportComputer"},CommonDBRelation:{name:"CommonDBRelation"},RuleImportComputerCollection:{name:"RuleImportComputerCollection"},RuleImportEntity:{name:"RuleImportEntity"},RuleImportEntityCollection:{name:"RuleImportEntityCollection"},RuleMailCollector:{name:"RuleMailCollector"},RuleMailCollectorCollection:{name:"RuleMailCollectorCollection"},RuleRight:{name:"RuleRight"},RuleRightCollection:{name:"RuleRightCollection"},RuleRightParameter:{name:"RuleRightParameter"},RuleSoftwareCategory:{name:"RuleSoftwareCategory"},RuleSoftwareCategoryCollection:{name:"RuleSoftwareCategoryCollection"},CommonDevice:{name:"CommonDevice"},RuleTicket:{name:"RuleTicket"},RuleTicketCollection:{name:"RuleTicketCollection"},SlaLevel:{name:"SlaLevel"},SlaLevelAction:{name:"SlaLevelAction"},SlaLevelCriteria:{name:"SlaLevelCriteria"},SoftwareCategory:{name:"SoftwareCategory"},SoftwareLicenseType:{name:"SoftwareLicenseType"},SoftwareVersion:{name:"SoftwareVersion"},SolutionTemplate:{name:"SolutionTemplate"},SolutionType:{name:"SolutionType"},CommonImplicitTreeDropdown:{name:"CommonImplicitTreeDropdown"},SsoVariable:{name:"SsoVariable"},State:{name:"State"},Supplier_Ticket:{name:"Supplier_Ticket"},SupplierType:{name:"SupplierType"},TaskCategory:{name:"TaskCategory"},Ticket:{name:"Ticket"},Ticket_Ticket:{name:"Ticket_Ticket"},Ticket_User:{name:"Ticket_User"},TicketCost:{name:"TicketCost"},TicketRecurrent:{name:"TicketRecurrent"},CommonITILActor:{name:"CommonITILActor"},TicketTask:{name:"TicketTask"},TicketTemplate:{name:"TicketTemplate"},TicketTemplateHiddenField:{name:"TicketTemplateHiddenField"},TicketTemplateMandatoryField:{name:"TicketTemplateMandatoryField"},TicketTemplatePredefinedField:{name:"TicketTemplatePredefinedField"},TicketValidation:{name:"TicketValidation"},UserCategory:{name:"UserCategory"},UserEmail:{name:"UserEmail"},UserTitle:{name:"UserTitle"},VirtualMachineState:{name:"VirtualMachineState"},CommonITILCost:{name:"CommonITILCost"},VirtualMachineSystem:{name:"VirtualMachineSystem"},VirtualMachineType:{name:"VirtualMachineType"},Vlan:{name:"Vlan"},WifiNetwork:{name:"WifiNetwork"},CommonITILValidation:{name:"CommonITILValidation"},CommonTreeDropdown:{name:"CommonTreeDropdown"},Bookmark_User:{name:"Bookmark_User"},Computer_Item:{name:"Computer_Item"},Computer_SoftwareLicense:{name:"Computer_SoftwareLicense"},Computer_SoftwareVersion:{name:"Computer_SoftwareVersion"},ComputerDisk:{name:"ComputerDisk"},ComputerModel:{name:"ComputerModel"},ComputerType:{name:"ComputerType"},ComputerVirtualMachine:{name:"ComputerVirtualMachine"},Consumable:{name:"Consumable"},ConsumableItemType:{name:"ConsumableItemType"},Contact_Supplier:{name:"Contact_Supplier"},Budget:{name:"Budget"},ContactType:{name:"ContactType"},Contract_Item:{name:"Contract_Item"},Contract_Supplier:{name:"Contract_Supplier"},ContractCost:{name:"ContractCost"},ContractType:{name:"ContractType"},DeviceCase:{name:"DeviceCase"},DeviceCaseType:{name:"DeviceCaseType"},DeviceControl:{name:"DeviceControl"},DeviceDrive:{name:"DeviceDrive"},DeviceGraphicCard:{name:"DeviceGraphicCard"},Calendar:{name:"Calendar"},DeviceHardDrive:{name:"DeviceHardDrive"},DeviceMemory:{name:"DeviceMemory"},DeviceMemoryType:{name:"DeviceMemoryType"},DeviceMotherboard:{name:"DeviceMotherboard"},DeviceNetworkCard:{name:"DeviceNetworkCard"},DevicePci:{name:"DevicePci"},DevicePowerSupply:{name:"DevicePowerSupply"},DeviceProcessor:{name:"DeviceProcessor"},DeviceSoundCard:{name:"DeviceSoundCard"},Document_Item:{name:"Document_Item"},Calendar_Holiday:{name:"Calendar_Holiday"},DocumentCategory:{name:"DocumentCategory"},DocumentType:{name:"DocumentType"},Domain:{name:"Domain"},DropdownTranslation:{name:"DropdownTranslation"},Entity:{name:"Entity"},Entity_KnowbaseItem:{name:"Entity_KnowbaseItem"},Entity_Reminder:{name:"Entity_Reminder"},Entity_RSSFeed:{name:"Entity_RSSFeed"},Fieldblacklist:{name:"Fieldblacklist"},FieldUnicity:{name:"FieldUnicity"},CalendarSegment:{name:"CalendarSegment"},Filesystem:{name:"Filesystem"},FQDN:{name:"FQDN"},FQDNLabel:{name:"FQDNLabel"},Group:{name:"Group"},Group_KnowbaseItem:{name:"Group_KnowbaseItem"},Group_Problem:{name:"Group_Problem"},Group_Reminder:{name:"Group_Reminder"},Group_RSSFeed:{name:"Group_RSSFeed"},Group_Ticket:{name:"Group_Ticket"},Group_User:{name:"Group_User"},Cartridge:{name:"Cartridge"},Holiday:{name:"Holiday"},Infocom:{name:"Infocom"},InterfaceType:{name:"InterfaceType"},IPAddress:{name:"IPAddress"},IPAddress_IPNetwork:{name:"IPAddress_IPNetwork"},IPNetmask:{name:"IPNetmask"},IPNetwork:{name:"IPNetwork"},IPNetwork_Vlan:{name:"IPNetwork_Vlan"},Item_DeviceCase:{name:"Item_DeviceCase"},Item_DeviceControl:{name:"Item_DeviceControl"},CartridgeItem_PrinterModel:{name:"CartridgeItem_PrinterModel"},Item_DeviceDrive:{name:"Item_DeviceDrive"},Item_DeviceGraphicCard:{name:"Item_DeviceGraphicCard"},Item_DeviceHardDrive:{name:"Item_DeviceHardDrive"},Item_DeviceMemory:{name:"Item_DeviceMemory"},Item_DeviceMotherboard:{name:"Item_DeviceMotherboard"},Item_DeviceNetworkCard:{name:"Item_DeviceNetworkCard"},Item_DevicePci:{name:"Item_DevicePci"},Item_DevicePowerSupply:{name:"Item_DevicePowerSupply"},Item_DeviceProcessor:{name:"Item_DeviceProcessor"},Item_Devices:{name:"Item_Devices"}}},function(e,t,r){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.itemtype=void 0;var n=i(r(2)),o=i(r(0));function i(e){return e&&e.__esModule?e:{default:e}}t.default=n.default,t.itemtype=o.default},function(e,t,r){"use strict";(function(t){var n=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e},o=function(){function e(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}return function(t,r,n){return r&&e(t.prototype,r),n&&e(t,n),t}}();var i=r(8),a=r(9).URL,s=r(0),u=function(){function e(t){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),this._url=t,this._sessionToken="",this._appToken=""}return o(e,[{key:"_prepareRequest",value:function(e,t,r,o){var s=new a(this._url);s.pathname+="/"+t;var u={"Content-Type":"application/json","Session-Token":this.sessionToken,"App-Token":this.appToken},c="";if(r&&(r.headers&&(u=n({},u,r.headers)),r["App-Token"]&&(this.appToken=r["App-Token"]),r.input&&(c=JSON.stringify(r),u=n({},u,{"Content-Length":c.length})),r.profiles_id&&(c=JSON.stringify(r),u=n({},u,{"Content-Length":c.length})),r.activeEntities&&(c=JSON.stringify(r.activeEntities),u=n({},u,{"Content-Length":c.length})),r.lostPassword&&(c=JSON.stringify(r.lostPassword),u=n({},u,{"Content-Length":c.length})),r.queryString))for(var l in r.queryString)r.queryString[l]&&s.searchParams.append(l,r.queryString[l]);var m={hostname:s.hostname,port:443,path:s.pathname+s.search,method:e,headers:u},f=i.request(m,function(e){var t="";e.on("data",function(e){t+=e}),e.on("end",function(){o(e,t)})});return f.on("error",function(){}),f.write(c),f}},{key:"_requestException",value:function(e,t){return{status:e,data:n({},t)}}},{key:"_assessStatus",value:function(e,t){if(e>=400&&e<600)return e>=400&&e<500?this._requestException(e,{codeError:JSON.parse(t)[0],messageError:JSON.parse(t)[1]}):this._requestException(e,JSON.parse(t))}},{key:"_isNumber",value:function(e){return!isNaN(e-0)&&null!==e&&""!==e&&!1!==e}},{key:"initSessionByCredentials",value:function(e,r,o){var i=this;return new Promise(function(a,s){try{var u={headers:{Authorization:"Basic "+t.from(e+":"+r).toString("base64")}};o&&(u.headers=n({},u.headers,{"App-Token":o})),i._prepareRequest("GET","initSession",u,function(e,t){var r=i._assessStatus(e.statusCode,t);r?s(r):(i.sessionToken=JSON.parse(t).session_token,a({status:e.statusCode,data:JSON.parse(t)}))}).end()}catch(e){s(e)}})}},{key:"initSessionByUserToken",value:function(e,t){var r=this;return new Promise(function(o,i){try{var a={headers:{Authorization:"user_token "+e}};t&&(a.headers=n({},a.headers,{"App-Token":t})),r._prepareRequest("GET","initSession",a,function(e,t){var n=r._assessStatus(e.statusCode,t);n?i(n):(r.sessionToken=JSON.parse(t).session_token,o({status:e.statusCode,data:JSON.parse(t)}))}).end()}catch(e){i(e)}})}},{key:"killSession",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","killSession",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):(e.sessionToken="",t({status:n.statusCode,data:{message:"User logout successfully"}}))}).end()}catch(e){r(e)}})}},{key:"getMyProfiles",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getMyProfiles",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o).myprofiles})}).end()}catch(e){r(e)}})}},{key:"getActiveProfile",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getActiveProfile",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o).active_profile})}).end()}catch(e){r(e)}})}},{key:"getMyEntities",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getMyEntities",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o).myentities})}).end()}catch(e){r(e)}})}},{key:"getActiveEntities",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getActiveEntities",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o)})}).end()}catch(e){r(e)}})}},{key:"getFullSession",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getFullSession",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o)})}).end()}catch(e){r(e)}})}},{key:"getGlpiConfig",value:function(){var e=this;return new Promise(function(t,r){try{e._prepareRequest("GET","getGlpiConfig",null,function(n,o){var i=e._assessStatus(n.statusCode,o);i?r(i):t({status:n.statusCode,data:JSON.parse(o)})}).end()}catch(e){r(e)}})}},{key:"getAllItems",value:function(e,t){var r=this;return new Promise(function(n,o){try{if(!e)throw r._requestException(400,"Invalid itemtype");if(e!==s[e.name])throw r._requestException(400,"Invalid itemtype");var i=null;t&&(i={queryString:t}),r._prepareRequest("GET",e.name,i,function(e,t){var i=r._assessStatus(e.statusCode,t);i?o(i):n({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){o(e)}})}},{key:"getAnItem",value:function(e,t,r){var n=this;return new Promise(function(o,i){try{if(!e)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(!n._isNumber(t))throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});var a=null;r&&(a={queryString:r});var u=e.name+"/"+t;n._prepareRequest("GET",u,a,function(e,t){var r=n._assessStatus(e.statusCode,t);r?i(r):o({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){i(e)}})}},{key:"getSubItems",value:function(e,t,r,n){var o=this;return new Promise(function(i,a){try{if(!e)throw o._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw o._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(!r)throw o._requestException(400,{errorCode:"",errorMessage:"Invalid sub itemtype"});if(r!==s[r.name])throw o._requestException(400,{errorCode:"",errorMessage:"Invalid sub itemtype"});if(!o._isNumber(t))throw o._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw o._requestException(400,{errorCode:"",errorMessage:"Invalid id"});var u=null;n&&(u={queryString:n});var c=e.name+"/"+t+"/"+r.name;o._prepareRequest("GET",c,u,function(e,t){var r=o._assessStatus(e.statusCode,t);r?a(r):i({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){a(e)}})}},{key:"addItem",value:function(e,t){var r=this;return new Promise(function(n,o){try{if(!e)throw r._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw r._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});var i={input:t},a=e.name;r._prepareRequest("POST",a,i,function(e,t){var i=r._assessStatus(e.statusCode,t);i?o(i):n({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){o(e)}})}},{key:"deleteItem",value:function(e,t,r,o){var i=this;return new Promise(function(a,u){try{if(!e)throw i._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw i._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});var c={};o&&(c={queryString:o});var l=e.name;if(t){if(!i._isNumber(t))throw i._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw i._requestException(400,{errorCode:"",errorMessage:"Invalid id"});l+="/"+t}else{if(!r)throw i._requestException(400,{errorCode:"",errorMessage:"Invalid input"});c=n({},c,{input:r})}i._prepareRequest("DELETE",l,c,function(e,t){var r=i._assessStatus(e.statusCode,t);r?u(r):a({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){u(e)}})}},{key:"updateItem",value:function(e,t,r){var n=this;return new Promise(function(o,i){try{if(!e)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});if(e!==s[e.name])throw n._requestException(400,{errorCode:"",errorMessage:"Invalid itemtype"});var a="",u=e.name;if(t){if(!n._isNumber(t))throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(t<0)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid id"});u+="/"+t}else{if(!r)throw n._requestException(400,{errorCode:"",errorMessage:"Invalid input"});a={input:r}}n._prepareRequest("PUT",u,a,function(e,t){var r=n._assessStatus(e.statusCode,t);r?i(r):o({status:e.statusCode,data:JSON.parse(t)})}).end()}catch(e){i(e)}})}},{key:"changeActiveProfile",value:function(e){var t=this;return new Promise(function(r,n){try{if(!t._isNumber(e))throw t._requestException(400,{errorCode:"",errorMessage:"Invalid id"});if(e<0)throw t._requestException(400,{errorCode:"",errorMessage:"Invalid id"});var o={profiles_id:e};t._prepareRequest("POTS","changeActiveProfile",o,function(e,o){var i=t._assessStatus(e.statusCode,o);i?n(i):r({status:e.statusCode,data:{message:"Active profile successfully changed"}})}).end()}catch(e){n(e)}})}},{key:"changeActiveEntities",value:function(e,t){var r=this;return new Promise(function(o,i){try{var a={activeEntities:{}};if(e){if(!r._isNumber(e))throw r._requestException(400,{errorCode:"",errorMessage:"Invalid entities_id"});if(e<0)throw r._requestException(400,{errorCode:"",errorMessage:"Invalid entities_id"});a={activeEntities:e}}if(t){if("boolean"!=typeof t)throw r._requestException(400,{errorCode:"",errorMessage:"Invalid is_recursive"});a={activeEntities:n({},a.activeEntities,{is_recursive:t})}}r._prepareRequest("POTS","changeActiveEntities",a,function(e,t){var n=r._assessStatus(e.statusCode,t);n?i(n):o(t?{status:e.statusCode,data:{message:"Active entities successfully changed"}}:{status:e.statusCode,data:{message:"Failed to change active entities"}})}).end()}catch(e){i(e)}})}},{key:"resetPasswordRequest",value:function(e){var t=this;return new Promise(function(r,n){try{var o={lostPassword:{email:e}};t._prepareRequest("PUT","lostPassword",o,function(e,o){var i=t._assessStatus(e.statusCode,o);i?n(i):r({status:e.statusCode,data:JSON.parse(o)[0]})}).end()}catch(e){n(e)}})}},{key:"passwordReset",value:function(e,t,r){var n=this;return new Promise(function(o,i){try{var a={lostPassword:{email:e,password_forget_token:t,password:r}};n._prepareRequest("PUT","lostPassword",a,function(e,t){var r=n._assessStatus(e.statusCode,t);r?i(r):o({status:e.statusCode,data:JSON.parse(t)[0]})}).end()}catch(e){i(e)}})}},{key:"url",get:function(){return this._url},set:function(e){this._url=e}},{key:"sessionToken",get:function(){return this._sessionToken},set:function(e){e&&(this._sessionToken=e)}},{key:"appToken",get:function(){return this._appToken},set:function(e){e&&(this._appToken=e)}}]),e}();e.exports=u}).call(t,r(3).Buffer)},function(e,t,r){"use strict";(function(e){var n=r(5),o=r(6),i=r(7);function a(){return u.TYPED_ARRAY_SUPPORT?2147483647:1073741823}function s(e,t){if(a()<t)throw new RangeError("Invalid typed array length");return u.TYPED_ARRAY_SUPPORT?(e=new Uint8Array(t)).__proto__=u.prototype:(null===e&&(e=new u(t)),e.length=t),e}function u(e,t,r){if(!(u.TYPED_ARRAY_SUPPORT||this instanceof u))return new u(e,t,r);if("number"==typeof e){if("string"==typeof t)throw new Error("If encoding is specified then the first argument must be a string");return m(this,e)}return c(this,e,t,r)}function c(e,t,r,n){if("number"==typeof t)throw new TypeError('"value" argument must not be a number');return"undefined"!=typeof ArrayBuffer&&t instanceof ArrayBuffer?function(e,t,r,n){if(t.byteLength,r<0||t.byteLength<r)throw new RangeError("'offset' is out of bounds");if(t.byteLength<r+(n||0))throw new RangeError("'length' is out of bounds");t=void 0===r&&void 0===n?new Uint8Array(t):void 0===n?new Uint8Array(t,r):new Uint8Array(t,r,n);u.TYPED_ARRAY_SUPPORT?(e=t).__proto__=u.prototype:e=f(e,t);return e}(e,t,r,n):"string"==typeof t?function(e,t,r){"string"==typeof r&&""!==r||(r="utf8");if(!u.isEncoding(r))throw new TypeError('"encoding" must be a valid string encoding');var n=0|h(t,r),o=(e=s(e,n)).write(t,r);o!==n&&(e=e.slice(0,o));return e}(e,t,r):function(e,t){if(u.isBuffer(t)){var r=0|p(t.length);return 0===(e=s(e,r)).length?e:(t.copy(e,0,0,r),e)}if(t){if("undefined"!=typeof ArrayBuffer&&t.buffer instanceof ArrayBuffer||"length"in t)return"number"!=typeof t.length||(n=t.length)!=n?s(e,0):f(e,t);if("Buffer"===t.type&&i(t.data))return f(e,t.data)}var n;throw new TypeError("First argument must be a string, Buffer, ArrayBuffer, Array, or array-like object.")}(e,t)}function l(e){if("number"!=typeof e)throw new TypeError('"size" argument must be a number');if(e<0)throw new RangeError('"size" argument must not be negative')}function m(e,t){if(l(t),e=s(e,t<0?0:0|p(t)),!u.TYPED_ARRAY_SUPPORT)for(var r=0;r<t;++r)e[r]=0;return e}function f(e,t){var r=t.length<0?0:0|p(t.length);e=s(e,r);for(var n=0;n<r;n+=1)e[n]=255&t[n];return e}function p(e){if(e>=a())throw new RangeError("Attempt to allocate Buffer larger than maximum size: 0x"+a().toString(16)+" bytes");return 0|e}function h(e,t){if(u.isBuffer(e))return e.length;if("undefined"!=typeof ArrayBuffer&&"function"==typeof ArrayBuffer.isView&&(ArrayBuffer.isView(e)||e instanceof ArrayBuffer))return e.byteLength;"string"!=typeof e&&(e=""+e);var r=e.length;if(0===r)return 0;for(var n=!1;;)switch(t){case"ascii":case"latin1":case"binary":return r;case"utf8":case"utf-8":case void 0:return B(e).length;case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return 2*r;case"hex":return r>>>1;case"base64":return L(e).length;default:if(n)return B(e).length;t=(""+t).toLowerCase(),n=!0}}function y(e,t,r){var n=e[t];e[t]=e[r],e[r]=n}function d(e,t,r,n,o){if(0===e.length)return-1;if("string"==typeof r?(n=r,r=0):r>2147483647?r=2147483647:r<-2147483648&&(r=-2147483648),r=+r,isNaN(r)&&(r=o?0:e.length-1),r<0&&(r=e.length+r),r>=e.length){if(o)return-1;r=e.length-1}else if(r<0){if(!o)return-1;r=0}if("string"==typeof t&&(t=u.from(t,n)),u.isBuffer(t))return 0===t.length?-1:g(e,t,r,n,o);if("number"==typeof t)return t&=255,u.TYPED_ARRAY_SUPPORT&&"function"==typeof Uint8Array.prototype.indexOf?o?Uint8Array.prototype.indexOf.call(e,t,r):Uint8Array.prototype.lastIndexOf.call(e,t,r):g(e,[t],r,n,o);throw new TypeError("val must be string, number or Buffer")}function g(e,t,r,n,o){var i,a=1,s=e.length,u=t.length;if(void 0!==n&&("ucs2"===(n=String(n).toLowerCase())||"ucs-2"===n||"utf16le"===n||"utf-16le"===n)){if(e.length<2||t.length<2)return-1;a=2,s/=2,u/=2,r/=2}function c(e,t){return 1===a?e[t]:e.readUInt16BE(t*a)}if(o){var l=-1;for(i=r;i<s;i++)if(c(e,i)===c(t,-1===l?0:i-l)){if(-1===l&&(l=i),i-l+1===u)return l*a}else-1!==l&&(i-=i-l),l=-1}else for(r+u>s&&(r=s-u),i=r;i>=0;i--){for(var m=!0,f=0;f<u;f++)if(c(e,i+f)!==c(t,f)){m=!1;break}if(m)return i}return-1}function C(e,t,r,n){r=Number(r)||0;var o=e.length-r;n?(n=Number(n))>o&&(n=o):n=o;var i=t.length;if(i%2!=0)throw new TypeError("Invalid hex string");n>i/2&&(n=i/2);for(var a=0;a<n;++a){var s=parseInt(t.substr(2*a,2),16);if(isNaN(s))return a;e[r+a]=s}return a}function v(e,t,r,n){return q(function(e){for(var t=[],r=0;r<e.length;++r)t.push(255&e.charCodeAt(r));return t}(t),e,r,n)}function w(e,t,r){return 0===t&&r===e.length?n.fromByteArray(e):n.fromByteArray(e.slice(t,r))}function T(e,t,r){r=Math.min(e.length,r);for(var n=[],o=t;o<r;){var i,a,s,u,c=e[o],l=null,m=c>239?4:c>223?3:c>191?2:1;if(o+m<=r)switch(m){case 1:c<128&&(l=c);break;case 2:128==(192&(i=e[o+1]))&&(u=(31&c)<<6|63&i)>127&&(l=u);break;case 3:i=e[o+1],a=e[o+2],128==(192&i)&&128==(192&a)&&(u=(15&c)<<12|(63&i)<<6|63&a)>2047&&(u<55296||u>57343)&&(l=u);break;case 4:i=e[o+1],a=e[o+2],s=e[o+3],128==(192&i)&&128==(192&a)&&128==(192&s)&&(u=(15&c)<<18|(63&i)<<12|(63&a)<<6|63&s)>65535&&u<1114112&&(l=u)}null===l?(l=65533,m=1):l>65535&&(l-=65536,n.push(l>>>10&1023|55296),l=56320|1023&l),n.push(l),o+=m}return function(e){var t=e.length;if(t<=_)return String.fromCharCode.apply(String,e);var r="",n=0;for(;n<t;)r+=String.fromCharCode.apply(String,e.slice(n,n+=_));return r}(n)}t.Buffer=u,t.SlowBuffer=function(e){+e!=e&&(e=0);return u.alloc(+e)},t.INSPECT_MAX_BYTES=50,u.TYPED_ARRAY_SUPPORT=void 0!==e.TYPED_ARRAY_SUPPORT?e.TYPED_ARRAY_SUPPORT:function(){try{var e=new Uint8Array(1);return e.__proto__={__proto__:Uint8Array.prototype,foo:function(){return 42}},42===e.foo()&&"function"==typeof e.subarray&&0===e.subarray(1,1).byteLength}catch(e){return!1}}(),t.kMaxLength=a(),u.poolSize=8192,u._augment=function(e){return e.__proto__=u.prototype,e},u.from=function(e,t,r){return c(null,e,t,r)},u.TYPED_ARRAY_SUPPORT&&(u.prototype.__proto__=Uint8Array.prototype,u.__proto__=Uint8Array,"undefined"!=typeof Symbol&&Symbol.species&&u[Symbol.species]===u&&Object.defineProperty(u,Symbol.species,{value:null,configurable:!0})),u.alloc=function(e,t,r){return n=null,i=t,a=r,l(o=e),o<=0?s(n,o):void 0!==i?"string"==typeof a?s(n,o).fill(i,a):s(n,o).fill(i):s(n,o);var n,o,i,a},u.allocUnsafe=function(e){return m(null,e)},u.allocUnsafeSlow=function(e){return m(null,e)},u.isBuffer=function(e){return!(null==e||!e._isBuffer)},u.compare=function(e,t){if(!u.isBuffer(e)||!u.isBuffer(t))throw new TypeError("Arguments must be Buffers");if(e===t)return 0;for(var r=e.length,n=t.length,o=0,i=Math.min(r,n);o<i;++o)if(e[o]!==t[o]){r=e[o],n=t[o];break}return r<n?-1:n<r?1:0},u.isEncoding=function(e){switch(String(e).toLowerCase()){case"hex":case"utf8":case"utf-8":case"ascii":case"latin1":case"binary":case"base64":case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return!0;default:return!1}},u.concat=function(e,t){if(!i(e))throw new TypeError('"list" argument must be an Array of Buffers');if(0===e.length)return u.alloc(0);var r;if(void 0===t)for(t=0,r=0;r<e.length;++r)t+=e[r].length;var n=u.allocUnsafe(t),o=0;for(r=0;r<e.length;++r){var a=e[r];if(!u.isBuffer(a))throw new TypeError('"list" argument must be an Array of Buffers');a.copy(n,o),o+=a.length}return n},u.byteLength=h,u.prototype._isBuffer=!0,u.prototype.swap16=function(){var e=this.length;if(e%2!=0)throw new RangeError("Buffer size must be a multiple of 16-bits");for(var t=0;t<e;t+=2)y(this,t,t+1);return this},u.prototype.swap32=function(){var e=this.length;if(e%4!=0)throw new RangeError("Buffer size must be a multiple of 32-bits");for(var t=0;t<e;t+=4)y(this,t,t+3),y(this,t+1,t+2);return this},u.prototype.swap64=function(){var e=this.length;if(e%8!=0)throw new RangeError("Buffer size must be a multiple of 64-bits");for(var t=0;t<e;t+=8)y(this,t,t+7),y(this,t+1,t+6),y(this,t+2,t+5),y(this,t+3,t+4);return this},u.prototype.toString=function(){var e=0|this.length;return 0===e?"":0===arguments.length?T(this,0,e):function(e,t,r){var n=!1;if((void 0===t||t<0)&&(t=0),t>this.length)return"";if((void 0===r||r>this.length)&&(r=this.length),r<=0)return"";if((r>>>=0)<=(t>>>=0))return"";for(e||(e="utf8");;)switch(e){case"hex":return S(this,t,r);case"utf8":case"utf-8":return T(this,t,r);case"ascii":return P(this,t,r);case"latin1":case"binary":return R(this,t,r);case"base64":return w(this,t,r);case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return k(this,t,r);default:if(n)throw new TypeError("Unknown encoding: "+e);e=(e+"").toLowerCase(),n=!0}}.apply(this,arguments)},u.prototype.equals=function(e){if(!u.isBuffer(e))throw new TypeError("Argument must be a Buffer");return this===e||0===u.compare(this,e)},u.prototype.inspect=function(){var e="",r=t.INSPECT_MAX_BYTES;return this.length>0&&(e=this.toString("hex",0,r).match(/.{2}/g).join(" "),this.length>r&&(e+=" ... ")),"<Buffer "+e+">"},u.prototype.compare=function(e,t,r,n,o){if(!u.isBuffer(e))throw new TypeError("Argument must be a Buffer");if(void 0===t&&(t=0),void 0===r&&(r=e?e.length:0),void 0===n&&(n=0),void 0===o&&(o=this.length),t<0||r>e.length||n<0||o>this.length)throw new RangeError("out of range index");if(n>=o&&t>=r)return 0;if(n>=o)return-1;if(t>=r)return 1;if(t>>>=0,r>>>=0,n>>>=0,o>>>=0,this===e)return 0;for(var i=o-n,a=r-t,s=Math.min(i,a),c=this.slice(n,o),l=e.slice(t,r),m=0;m<s;++m)if(c[m]!==l[m]){i=c[m],a=l[m];break}return i<a?-1:a<i?1:0},u.prototype.includes=function(e,t,r){return-1!==this.indexOf(e,t,r)},u.prototype.indexOf=function(e,t,r){return d(this,e,t,r,!0)},u.prototype.lastIndexOf=function(e,t,r){return d(this,e,t,r,!1)},u.prototype.write=function(e,t,r,n){if(void 0===t)n="utf8",r=this.length,t=0;else if(void 0===r&&"string"==typeof t)n=t,r=this.length,t=0;else{if(!isFinite(t))throw new Error("Buffer.write(string, encoding, offset[, length]) is no longer supported");t|=0,isFinite(r)?(r|=0,void 0===n&&(n="utf8")):(n=r,r=void 0)}var o=this.length-t;if((void 0===r||r>o)&&(r=o),e.length>0&&(r<0||t<0)||t>this.length)throw new RangeError("Attempt to write outside buffer bounds");n||(n="utf8");for(var i,a,s,u,c,l,m,f,p,h=!1;;)switch(n){case"hex":return C(this,e,t,r);case"utf8":case"utf-8":return f=t,p=r,q(B(e,(m=this).length-f),m,f,p);case"ascii":return v(this,e,t,r);case"latin1":case"binary":return v(this,e,t,r);case"base64":return u=this,c=t,l=r,q(L(e),u,c,l);case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return a=t,s=r,q(function(e,t){for(var r,n,o,i=[],a=0;a<e.length&&!((t-=2)<0);++a)r=e.charCodeAt(a),n=r>>8,o=r%256,i.push(o),i.push(n);return i}(e,(i=this).length-a),i,a,s);default:if(h)throw new TypeError("Unknown encoding: "+n);n=(""+n).toLowerCase(),h=!0}},u.prototype.toJSON=function(){return{type:"Buffer",data:Array.prototype.slice.call(this._arr||this,0)}};var _=4096;function P(e,t,r){var n="";r=Math.min(e.length,r);for(var o=t;o<r;++o)n+=String.fromCharCode(127&e[o]);return n}function R(e,t,r){var n="";r=Math.min(e.length,r);for(var o=t;o<r;++o)n+=String.fromCharCode(e[o]);return n}function S(e,t,r){var n=e.length;(!t||t<0)&&(t=0),(!r||r<0||r>n)&&(r=n);for(var o="",i=t;i<r;++i)o+=O(e[i]);return o}function k(e,t,r){for(var n=e.slice(t,r),o="",i=0;i<n.length;i+=2)o+=String.fromCharCode(n[i]+256*n[i+1]);return o}function D(e,t,r){if(e%1!=0||e<0)throw new RangeError("offset is not uint");if(e+t>r)throw new RangeError("Trying to access beyond buffer length")}function I(e,t,r,n,o,i){if(!u.isBuffer(e))throw new TypeError('"buffer" argument must be a Buffer instance');if(t>o||t<i)throw new RangeError('"value" argument is out of bounds');if(r+n>e.length)throw new RangeError("Index out of range")}function E(e,t,r,n){t<0&&(t=65535+t+1);for(var o=0,i=Math.min(e.length-r,2);o<i;++o)e[r+o]=(t&255<<8*(n?o:1-o))>>>8*(n?o:1-o)}function b(e,t,r,n){t<0&&(t=4294967295+t+1);for(var o=0,i=Math.min(e.length-r,4);o<i;++o)e[r+o]=t>>>8*(n?o:3-o)&255}function N(e,t,r,n,o,i){if(r+n>e.length)throw new RangeError("Index out of range");if(r<0)throw new RangeError("Index out of range")}function M(e,t,r,n,i){return i||N(e,0,r,4),o.write(e,t,r,n,23,4),r+4}function A(e,t,r,n,i){return i||N(e,0,r,8),o.write(e,t,r,n,52,8),r+8}u.prototype.slice=function(e,t){var r,n=this.length;if(e=~~e,t=void 0===t?n:~~t,e<0?(e+=n)<0&&(e=0):e>n&&(e=n),t<0?(t+=n)<0&&(t=0):t>n&&(t=n),t<e&&(t=e),u.TYPED_ARRAY_SUPPORT)(r=this.subarray(e,t)).__proto__=u.prototype;else{var o=t-e;r=new u(o,void 0);for(var i=0;i<o;++i)r[i]=this[i+e]}return r},u.prototype.readUIntLE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=this[e],o=1,i=0;++i<t&&(o*=256);)n+=this[e+i]*o;return n},u.prototype.readUIntBE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=this[e+--t],o=1;t>0&&(o*=256);)n+=this[e+--t]*o;return n},u.prototype.readUInt8=function(e,t){return t||D(e,1,this.length),this[e]},u.prototype.readUInt16LE=function(e,t){return t||D(e,2,this.length),this[e]|this[e+1]<<8},u.prototype.readUInt16BE=function(e,t){return t||D(e,2,this.length),this[e]<<8|this[e+1]},u.prototype.readUInt32LE=function(e,t){return t||D(e,4,this.length),(this[e]|this[e+1]<<8|this[e+2]<<16)+16777216*this[e+3]},u.prototype.readUInt32BE=function(e,t){return t||D(e,4,this.length),16777216*this[e]+(this[e+1]<<16|this[e+2]<<8|this[e+3])},u.prototype.readIntLE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=this[e],o=1,i=0;++i<t&&(o*=256);)n+=this[e+i]*o;return n>=(o*=128)&&(n-=Math.pow(2,8*t)),n},u.prototype.readIntBE=function(e,t,r){e|=0,t|=0,r||D(e,t,this.length);for(var n=t,o=1,i=this[e+--n];n>0&&(o*=256);)i+=this[e+--n]*o;return i>=(o*=128)&&(i-=Math.pow(2,8*t)),i},u.prototype.readInt8=function(e,t){return t||D(e,1,this.length),128&this[e]?-1*(255-this[e]+1):this[e]},u.prototype.readInt16LE=function(e,t){t||D(e,2,this.length);var r=this[e]|this[e+1]<<8;return 32768&r?4294901760|r:r},u.prototype.readInt16BE=function(e,t){t||D(e,2,this.length);var r=this[e+1]|this[e]<<8;return 32768&r?4294901760|r:r},u.prototype.readInt32LE=function(e,t){return t||D(e,4,this.length),this[e]|this[e+1]<<8|this[e+2]<<16|this[e+3]<<24},u.prototype.readInt32BE=function(e,t){return t||D(e,4,this.length),this[e]<<24|this[e+1]<<16|this[e+2]<<8|this[e+3]},u.prototype.readFloatLE=function(e,t){return t||D(e,4,this.length),o.read(this,e,!0,23,4)},u.prototype.readFloatBE=function(e,t){return t||D(e,4,this.length),o.read(this,e,!1,23,4)},u.prototype.readDoubleLE=function(e,t){return t||D(e,8,this.length),o.read(this,e,!0,52,8)},u.prototype.readDoubleBE=function(e,t){return t||D(e,8,this.length),o.read(this,e,!1,52,8)},u.prototype.writeUIntLE=function(e,t,r,n){(e=+e,t|=0,r|=0,n)||I(this,e,t,r,Math.pow(2,8*r)-1,0);var o=1,i=0;for(this[t]=255&e;++i<r&&(o*=256);)this[t+i]=e/o&255;return t+r},u.prototype.writeUIntBE=function(e,t,r,n){(e=+e,t|=0,r|=0,n)||I(this,e,t,r,Math.pow(2,8*r)-1,0);var o=r-1,i=1;for(this[t+o]=255&e;--o>=0&&(i*=256);)this[t+o]=e/i&255;return t+r},u.prototype.writeUInt8=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,1,255,0),u.TYPED_ARRAY_SUPPORT||(e=Math.floor(e)),this[t]=255&e,t+1},u.prototype.writeUInt16LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,65535,0),u.TYPED_ARRAY_SUPPORT?(this[t]=255&e,this[t+1]=e>>>8):E(this,e,t,!0),t+2},u.prototype.writeUInt16BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,65535,0),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>8,this[t+1]=255&e):E(this,e,t,!1),t+2},u.prototype.writeUInt32LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,4294967295,0),u.TYPED_ARRAY_SUPPORT?(this[t+3]=e>>>24,this[t+2]=e>>>16,this[t+1]=e>>>8,this[t]=255&e):b(this,e,t,!0),t+4},u.prototype.writeUInt32BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,4294967295,0),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>24,this[t+1]=e>>>16,this[t+2]=e>>>8,this[t+3]=255&e):b(this,e,t,!1),t+4},u.prototype.writeIntLE=function(e,t,r,n){if(e=+e,t|=0,!n){var o=Math.pow(2,8*r-1);I(this,e,t,r,o-1,-o)}var i=0,a=1,s=0;for(this[t]=255&e;++i<r&&(a*=256);)e<0&&0===s&&0!==this[t+i-1]&&(s=1),this[t+i]=(e/a>>0)-s&255;return t+r},u.prototype.writeIntBE=function(e,t,r,n){if(e=+e,t|=0,!n){var o=Math.pow(2,8*r-1);I(this,e,t,r,o-1,-o)}var i=r-1,a=1,s=0;for(this[t+i]=255&e;--i>=0&&(a*=256);)e<0&&0===s&&0!==this[t+i+1]&&(s=1),this[t+i]=(e/a>>0)-s&255;return t+r},u.prototype.writeInt8=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,1,127,-128),u.TYPED_ARRAY_SUPPORT||(e=Math.floor(e)),e<0&&(e=255+e+1),this[t]=255&e,t+1},u.prototype.writeInt16LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,32767,-32768),u.TYPED_ARRAY_SUPPORT?(this[t]=255&e,this[t+1]=e>>>8):E(this,e,t,!0),t+2},u.prototype.writeInt16BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,2,32767,-32768),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>8,this[t+1]=255&e):E(this,e,t,!1),t+2},u.prototype.writeInt32LE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,2147483647,-2147483648),u.TYPED_ARRAY_SUPPORT?(this[t]=255&e,this[t+1]=e>>>8,this[t+2]=e>>>16,this[t+3]=e>>>24):b(this,e,t,!0),t+4},u.prototype.writeInt32BE=function(e,t,r){return e=+e,t|=0,r||I(this,e,t,4,2147483647,-2147483648),e<0&&(e=4294967295+e+1),u.TYPED_ARRAY_SUPPORT?(this[t]=e>>>24,this[t+1]=e>>>16,this[t+2]=e>>>8,this[t+3]=255&e):b(this,e,t,!1),t+4},u.prototype.writeFloatLE=function(e,t,r){return M(this,e,t,!0,r)},u.prototype.writeFloatBE=function(e,t,r){return M(this,e,t,!1,r)},u.prototype.writeDoubleLE=function(e,t,r){return A(this,e,t,!0,r)},u.prototype.writeDoubleBE=function(e,t,r){return A(this,e,t,!1,r)},u.prototype.copy=function(e,t,r,n){if(r||(r=0),n||0===n||(n=this.length),t>=e.length&&(t=e.length),t||(t=0),n>0&&n<r&&(n=r),n===r)return 0;if(0===e.length||0===this.length)return 0;if(t<0)throw new RangeError("targetStart out of bounds");if(r<0||r>=this.length)throw new RangeError("sourceStart out of bounds");if(n<0)throw new RangeError("sourceEnd out of bounds");n>this.length&&(n=this.length),e.length-t<n-r&&(n=e.length-t+r);var o,i=n-r;if(this===e&&r<t&&t<n)for(o=i-1;o>=0;--o)e[o+t]=this[o+r];else if(i<1e3||!u.TYPED_ARRAY_SUPPORT)for(o=0;o<i;++o)e[o+t]=this[o+r];else Uint8Array.prototype.set.call(e,this.subarray(r,r+i),t);return i},u.prototype.fill=function(e,t,r,n){if("string"==typeof e){if("string"==typeof t?(n=t,t=0,r=this.length):"string"==typeof r&&(n=r,r=this.length),1===e.length){var o=e.charCodeAt(0);o<256&&(e=o)}if(void 0!==n&&"string"!=typeof n)throw new TypeError("encoding must be a string");if("string"==typeof n&&!u.isEncoding(n))throw new TypeError("Unknown encoding: "+n)}else"number"==typeof e&&(e&=255);if(t<0||this.length<t||this.length<r)throw new RangeError("Out of range index");if(r<=t)return this;var i;if(t>>>=0,r=void 0===r?this.length:r>>>0,e||(e=0),"number"==typeof e)for(i=t;i<r;++i)this[i]=e;else{var a=u.isBuffer(e)?e:B(new u(e,n).toString()),s=a.length;for(i=0;i<r-t;++i)this[i+t]=a[i%s]}return this};var U=/[^+\/0-9A-Za-z-_]/g;function O(e){return e<16?"0"+e.toString(16):e.toString(16)}function B(e,t){var r;t=t||1/0;for(var n=e.length,o=null,i=[],a=0;a<n;++a){if((r=e.charCodeAt(a))>55295&&r<57344){if(!o){if(r>56319){(t-=3)>-1&&i.push(239,191,189);continue}if(a+1===n){(t-=3)>-1&&i.push(239,191,189);continue}o=r;continue}if(r<56320){(t-=3)>-1&&i.push(239,191,189),o=r;continue}r=65536+(o-55296<<10|r-56320)}else o&&(t-=3)>-1&&i.push(239,191,189);if(o=null,r<128){if((t-=1)<0)break;i.push(r)}else if(r<2048){if((t-=2)<0)break;i.push(r>>6|192,63&r|128)}else if(r<65536){if((t-=3)<0)break;i.push(r>>12|224,r>>6&63|128,63&r|128)}else{if(!(r<1114112))throw new Error("Invalid code point");if((t-=4)<0)break;i.push(r>>18|240,r>>12&63|128,r>>6&63|128,63&r|128)}}return i}function L(e){return n.toByteArray(function(e){var t;if((e=(t=e,t.trim?t.trim():t.replace(/^\s+|\s+$/g,"")).replace(U,"")).length<2)return"";for(;e.length%4!=0;)e+="=";return e}(e))}function q(e,t,r,n){for(var o=0;o<n&&!(o+r>=t.length||o>=e.length);++o)t[o+r]=e[o];return o}}).call(t,r(4))},function(e,t){var r;r=function(){return this}();try{r=r||Function("return this")()||(0,eval)("this")}catch(e){"object"==typeof window&&(r=window)}e.exports=r},function(e,t,r){"use strict";t.byteLength=function(e){return 3*e.length/4-c(e)},t.toByteArray=function(e){var t,r,n,a,s,u=e.length;a=c(e),s=new i(3*u/4-a),r=a>0?u-4:u;var l=0;for(t=0;t<r;t+=4)n=o[e.charCodeAt(t)]<<18|o[e.charCodeAt(t+1)]<<12|o[e.charCodeAt(t+2)]<<6|o[e.charCodeAt(t+3)],s[l++]=n>>16&255,s[l++]=n>>8&255,s[l++]=255&n;2===a?(n=o[e.charCodeAt(t)]<<2|o[e.charCodeAt(t+1)]>>4,s[l++]=255&n):1===a&&(n=o[e.charCodeAt(t)]<<10|o[e.charCodeAt(t+1)]<<4|o[e.charCodeAt(t+2)]>>2,s[l++]=n>>8&255,s[l++]=255&n);return s},t.fromByteArray=function(e){for(var t,r=e.length,o=r%3,i="",a=[],s=0,u=r-o;s<u;s+=16383)a.push(l(e,s,s+16383>u?u:s+16383));1===o?(t=e[r-1],i+=n[t>>2],i+=n[t<<4&63],i+="=="):2===o&&(t=(e[r-2]<<8)+e[r-1],i+=n[t>>10],i+=n[t>>4&63],i+=n[t<<2&63],i+="=");return a.push(i),a.join("")};for(var n=[],o=[],i="undefined"!=typeof Uint8Array?Uint8Array:Array,a="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",s=0,u=a.length;s<u;++s)n[s]=a[s],o[a.charCodeAt(s)]=s;function c(e){var t=e.length;if(t%4>0)throw new Error("Invalid string. Length must be a multiple of 4");return"="===e[t-2]?2:"="===e[t-1]?1:0}function l(e,t,r){for(var o,i,a=[],s=t;s<r;s+=3)o=(e[s]<<16)+(e[s+1]<<8)+e[s+2],a.push(n[(i=o)>>18&63]+n[i>>12&63]+n[i>>6&63]+n[63&i]);return a.join("")}o["-".charCodeAt(0)]=62,o["_".charCodeAt(0)]=63},function(e,t){t.read=function(e,t,r,n,o){var i,a,s=8*o-n-1,u=(1<<s)-1,c=u>>1,l=-7,m=r?o-1:0,f=r?-1:1,p=e[t+m];for(m+=f,i=p&(1<<-l)-1,p>>=-l,l+=s;l>0;i=256*i+e[t+m],m+=f,l-=8);for(a=i&(1<<-l)-1,i>>=-l,l+=n;l>0;a=256*a+e[t+m],m+=f,l-=8);if(0===i)i=1-c;else{if(i===u)return a?NaN:1/0*(p?-1:1);a+=Math.pow(2,n),i-=c}return(p?-1:1)*a*Math.pow(2,i-n)},t.write=function(e,t,r,n,o,i){var a,s,u,c=8*i-o-1,l=(1<<c)-1,m=l>>1,f=23===o?Math.pow(2,-24)-Math.pow(2,-77):0,p=n?0:i-1,h=n?1:-1,y=t<0||0===t&&1/t<0?1:0;for(t=Math.abs(t),isNaN(t)||t===1/0?(s=isNaN(t)?1:0,a=l):(a=Math.floor(Math.log(t)/Math.LN2),t*(u=Math.pow(2,-a))<1&&(a--,u*=2),(t+=a+m>=1?f/u:f*Math.pow(2,1-m))*u>=2&&(a++,u/=2),a+m>=l?(s=0,a=l):a+m>=1?(s=(t*u-1)*Math.pow(2,o),a+=m):(s=t*Math.pow(2,m-1)*Math.pow(2,o),a=0));o>=8;e[r+p]=255&s,p+=h,s/=256,o-=8);for(a=a<<o|s,c+=o;c>0;e[r+p]=255&a,p+=h,a/=256,c-=8);e[r+p-h]|=128*y}},function(e,t){var r={}.toString;e.exports=Array.isArray||function(e){return"[object Array]"==r.call(e)}},function(t,r){t.exports=e},function(e,r){e.exports=t}])});
